% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{actionBtnClass}
\alias{actionBtnClass}
\title{Css style defining the action buttons}
\format{
An object of class \code{character} of length 1.
}
\usage{
actionBtnClass
}
\description{
Css style defining the action buttons
}
\section{Fields}{

\describe{
\item{\code{actionBtnClass}}{xxxx}
}}

\keyword{datasets}
