% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pipeline.R
\name{ActionOn_Data_Trigger}
\alias{ActionOn_Data_Trigger}
\title{xxx}
\usage{
ActionOn_Data_Trigger()
}
\value{
Nothing
}
\description{
Catch the return value of a module and update the list of isDone modules
This list is updated with the names of datasets present in the rv$tmp
variable. One set to TRUE all the elements in isDone which have a corresponding
element in names(rv$tmp).
One cannot simply set to TRUE the last element of rv$tmp because it will does
not work in case of a reseted module (it is not in the names(rv$tmp) list
anymore)
If a value (not NULL) is received, then it corresponds to the module
pointed by the current position
This function also updates the list isDone
This function updates the current dataset (self$rv$dataIn)
}
