% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R
\name{Add_Item_to_Dataset}
\alias{Add_Item_to_Dataset}
\title{xxx}
\usage{
Add_Item_to_Dataset(dataset, name)
}
\value{
The dataset minus some items
}
\description{
Adds one or more items to the dataset. This function is specific of the
type of dataset.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Add_Item_to_Dataset(Exp1_R25_pept, name = 'foo')

}
