% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{btn_success_color}
\alias{btn_success_color}
\title{Css class defining the success state}
\format{
An object of class \code{character} of length 1.
}
\usage{
btn_success_color
}
\description{
Css class defining the success state
}
\section{Fields}{

\describe{
\item{\code{btn_success_color}}{xxxx}
}}

\keyword{datasets}
