% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{Build_EncapsulateScreens_ui}
\alias{Build_EncapsulateScreens_ui}
\title{xxx}
\usage{
Build_EncapsulateScreens_ui(ns, id, steps, ll.UI)
}
\arguments{
\item{ns}{xxx}

\item{id}{xxx}

\item{steps}{xxx}

\item{ll.UI}{xxx}
}
\value{
A `renderUI` function
}
\description{
xxx
}
