% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_dataset.R
\name{Build_example_dataset}
\alias{Build_example_dataset}
\title{Example dataset builder for examples in this package.}
\usage{
Build_example_dataset()
}
\description{
xxx
}
\examples{
Build_example_dataset()

}
