% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nav_ui.R
\name{Build_nav_h_ui}
\alias{Build_nav_h_ui}
\title{Horizontal timeline UI}
\usage{
Build_nav_h_ui(ns)
}
\arguments{
\item{ns}{The namespace variable from the module that called this function.}
}
\value{
A shiny UI
}
\description{
This function displays the timeline UI with a horizontal layout
}
\author{
Samuel Wieczorek
}
