% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nav_ui.R
\name{Build_nav_v_ui}
\alias{Build_nav_v_ui}
\title{Vertical timeline UI}
\usage{
Build_nav_v_ui(ns)
}
\arguments{
\item{ns}{The namespace variable from the module that called this function.}
}
\value{
A shiny UI
}
\description{
This function displays the timeline UI with a vertical layout
}
\examples{
NULL


}
\author{
Samuel Wieczorek
}
