% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_nav_ui.R
\name{Build_nav_X_ui}
\alias{Build_nav_X_ui}
\alias{DisplayWholeUI}
\alias{Build_nav_v_ui}
\alias{Build_nav_h_ui}
\title{Timeline UI}
\usage{
DisplayWholeUI(ns, layout = "h")

Build_nav_v_ui(ns)

Build_nav_h_ui(ns)
}
\arguments{
\item{ns}{The namespace variable from the module that called this function.}
}
\value{
A shiny UI
}
\description{
These functions display the timeline UI with a vertical nor 
horizontal layout
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
