% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_build_report.R
\name{build_report}
\alias{build_report}
\alias{build_report_ui}
\alias{build_report_server}
\title{dl}
\usage{
build_report_ui(id)

build_report_server(
  id,
  dataIn = reactive({
     NULL
 }),
  filename = "myDataset"
)

build_report(data, filename = "myDataset")
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{filename}{xxx}

\item{data}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(build_report(lldata))

    shiny::runApp(build_report(lldata, filename = "myDataset"))
}

NULL
NULL
NULL

}
