% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{Build_SkippedInfoPanel}
\alias{Build_SkippedInfoPanel}
\title{xxx}
\usage{
Build_SkippedInfoPanel(steps.status, current.pos, config)
}
\arguments{
\item{steps.status}{xxx}

\item{current.pos}{xxx}

\item{config}{xxx}
}
\value{
A `wellPanel`
}
\description{
xxx
}
