% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{BuildData2Send}
\alias{BuildData2Send}
\title{xxx}
\usage{
BuildData2Send(dataIn, stepsNames)
}
\arguments{
\item{dataIn}{An instance of the `QFeatures` class}

\item{stepsNames}{xxx}
}
\value{
A vector of QF datasets
}
\description{
xxx
}
\examples{
NULL
}
