% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{call_func}
\alias{call_func}
\title{Wrapper to the function `do.call`}
\usage{
call_func(fname, args)
}
\arguments{
\item{fname}{The name of the function to execute}

\item{args}{The `list()` of its arguments}
}
\value{
The result of the function called
}
\description{
Wrapper to the function `do.call`
}
\examples{
call_func("stats::rnorm", list(n = 10, mean = 3))

}
\seealso{
[do.call()]
}
