% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{CheckConfig}
\alias{CheckConfig}
\title{xxx}
\usage{
CheckConfig(config)
}
\arguments{
\item{config}{A list containing the rv$configuration of the current object.
See xxx}
}
\value{
A list of two items:
* `passed`: A boolean that indicates if the config is correct or not.
* `msg`: A `character(1)` as message.
}
\description{
xxx

Check if the rv$config is correct
}
\examples{
conf <- list(mode = "process",
steps = c("Step 1", "Step 2", "Save"),
mandatory = c( TRUE, FALSE, TRUE),
path_to_md_dir = system.file('module_examples/md/', package='Magellan')
)
CheckConfig(conf)

}
\author{
Samuel Wieczorek
}
