% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{CheckProcessCode}
\alias{CheckProcessCode}
\title{Check source code of a module process}
\usage{
CheckProcessCode(sourcefile = NULL)
}
\arguments{
\item{sourcefile}{xxx}
}
\value{
A list of two items:
* `passed`: A boolean that indicates if the config is correct or not.
* `msg`: A `character(1)` as message.
}
\description{
Check if the source code of a process module is correct
}
\details{
xxxx
xxx
xxxx
}
\author{
Samuel Wieczorek
}
