% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckWorkflowDir}
\alias{CheckWorkflowDir}
\title{Basic check workflow directory}
\usage{
CheckWorkflowDir(path)
}
\arguments{
\item{path}{A `character(1)`}
}
\value{
A `boolean(1)`
}
\description{
This function checks if the directory contains well-formed directories and files
It must contains 3 directories: 'md', 'R' and 'data'. 
The 'R' directory must contains two directories:
* 'workflows' that contains the source files for workflows,
* 'other' that contains additional source files used by workflows. This directory 
can be empty. For each
file in the 'R/workflows' directory, there must exists a *.md file with the same filename
in the 'md' directory.
The 'data' directory can be empty.

For a full description of the nomenclature of workflows filename, please refer
to xxx.
}
