% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir.darwin}
\alias{choose.dir.darwin}
\title{The apple version of the choose folder}
\usage{
choose.dir.darwin(default = NA, caption = NA)

choose.dir.darwin(default = NA, caption = NA)
}
\value{
A length one character vector, character NA if 'Cancel' was selected.

A length one character vector, character NA if 'Cancel' was selected.
}
\description{
The apple version of the choose folder

The apple version of the choose folder
}
\seealso{
\code{\link{choose.dir}}

\code{\link{choose.dir}}
}
