% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{CleanSourceCode}
\alias{CleanSourceCode}
\title{Clean source code before syntax analysis
xxx}
\usage{
CleanSourceCode(source = NULL)
}
\arguments{
\item{source}{A vector in which each element is a line read from source code file}
}
\value{
NA
}
\description{
xxx

Check if the rv$config is correct
}
\examples{
conf <- list(parent = "pipeline",
name = "process",
steps = c('Description', "Step 1", "Step 2", "Save"),
mandatory = c(TRUE, TRUE, FALSE, TRUE)
)


}
\author{
Samuel Wieczorek
}
