% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclip.R
\name{clipboard}
\alias{clipboard}
\title{xxx}
\usage{
clipboard(x, sep = "\\t", row.names = FALSE, col.names = F)
}
\description{
From: https://stackoverflow.com/questions/56942384/error-in-clipboard-on-x11-requires-that-the-display-envvar-be-configured
}
