% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Code_ObserveEvent_ValidationBtns}
\alias{Code_ObserveEvent_ValidationBtns}
\title{Code for xxxx}
\usage{
Code_ObserveEvent_ValidationBtns()
}
\description{
xxxx
}
\examples{
\dontrun{
code <- Code_ObserveEvent_ValidationBtns()
cat(code)
}

}
\author{
Samuel Wieczorek
}
