% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\name{Config}
\alias{Config}
\alias{initialize,Config-method}
\title{Config class definition}
\usage{
xxx

\S4method{initialize}{Config}(
  .Object,
  name,
  parent,
  mode,
  steps,
  mandatory,
  path_to_md_file,
  path
)

Config(
  name,
  parent = "",
  mode,
  steps,
  mandatory,
  path_to_md_file = NULL,
  path = NULL
)
}
\arguments{
\item{name}{xxx}

\item{parent}{xxx}

\item{mode}{xxx}

\item{steps}{xxx}

\item{mandatory}{xxx}

\item{path_to_md_file}{xxx}

\item{path}{xxx}

\item{.Object}{xxx}

\item{object}{xxx}
}
\description{
This class is used to store the configuration of any process
used with MagellanNTK It contains a validity function to ensure
that the format is correct.

Validity:
* The first step must be called 'Description', it is a mandatory step. Thus, 
the first item of the mandatory vector is TRUE.
To be continued...

Wrapper function to the constructor of the class
}
\section{Slots}{

\describe{
\item{\code{name}}{xxx}

\item{\code{parent}}{xxx}

\item{\code{mode}}{xxx}

\item{\code{steps}}{xxx}

\item{\code{mandatory}}{xxx}

\item{\code{path_to_md_file}}{xxx}

\item{\code{path}}{xxx}

\item{\code{ll.UI}}{xxx}

\item{\code{steps.source.file}}{xxx}
}}

\examples{
config <- Config(
mode = 'pipeline',
name = 'PipelineA',
# List of all steps of the process
# Here, each step is a workflow
steps = c('Description', 'Process1', 'Process2', 'Process3'),
# A vector of boolean indicating if the steps are mandatory or not.
mandatory = c(TRUE, FALSE, FALSE, TRUE),

path_to_md_file = system.file('extdata/module_examples/md/PipelineA.md', package='MagellanNTK')
)

}
