% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraFunctions.R
\name{create_custom_funcs}
\alias{create_custom_funcs}
\alias{default_add_func}
\alias{default_keep_func}
\alias{createExtraFunctions}
\title{xxx}
\usage{
default_add_func()

default_keep_func()

createExtraFunctions(
  path.dir = ".",
  add_func = default_add_func(),
  keep_func = default_keep_func()
)
}
\arguments{
\item{path.dir}{xxx}

\item{add_func}{xxx}

\item{keep_func}{xxx}

\item{dataset}{xxx}

\item{name}{the name to associate to the dataset in the object list}

\item{object}{An instance of type list. Must get TRUE to inherits(object, 'list')}

\item{range}{xxx}
}
\description{
xxx

This function appends a dataset in the list with customization

This function deletes the items not included in the
}
