% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraModules.R
\name{create_extra_module}
\alias{create_extra_module}
\alias{createExtraModule}
\alias{extra_module}
\alias{write_extraModule_code}
\title{Creates the template extar module}
\usage{
createExtraModule(name, path = ".")

write_extraModule_code(con, name)
}
\arguments{
\item{name}{The name}

\item{path}{xxx}

\item{con}{xxx}

\item{id}{xxx}
}
\description{
Creates an exemplary extra module to be included into
a step. 
Int the code that will be produced, the parts that are not commented 
are mandatory ans should not be modified.
The commented parts are just for example purpose and need to be
customize to your function

## To be customized ##
}
\examples{
if (interactive()){
data(ft_na)
ui <- foo_ui('query')

server <- function(input, output, session) {
  
  rv <- reactiveValues(
    res = NULL
  )
  ll.tags <- c('None' = 'None', 
  qMetadata.def(typeDataset(ft_na[[1]]))$node)
  
  rv$res <- foo_server('query', 
  reset = reactive({NULL}),
  is.enabled = reactive({NULL})
  )
  
  observeEvent(rv$res$dataOut()$value, ignoreNULL = TRUE, ignoreInit = TRUE, {
    print(rv$res$dataOut()$value)
    print(rv$res$dataOut()$widgets)
  })
}

shinyApp(ui=ui, server=server)
}
}
