% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMdDescriptionFiles.R
\name{Create_md_file}
\alias{Create_md_file}
\title{xxx}
\usage{
Create_md_file(path, fullname)
}
\arguments{
\item{path}{xxx}

\item{fullname}{xxx}

\item{mode}{xxx}

\item{rawText}{xxx}
}
\description{
xxx
}
