% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModuleTemplate.R
\name{createModuleTemplate}
\alias{createModuleTemplate}
\title{Create process template code}
\usage{
createModuleTemplate(name = NULL, config = NULL)
}
\arguments{
\item{name}{xxx}

\item{config}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
conf.process <- list(
    mode = "process",
    steps = c("Step 1", "Step 2", "Save"),
    mandatory = c(TRUE, FALSE, TRUE),
    path_to_md_dir = NULL
)
createModuleTemplate("Process1", conf.process)

conf.pipeline <- list(
    mode = "pipeline",
    steps = c("Process 1", "Process 2", "Process 3"),
    mandatory = c(TRUE, FALSE, TRUE),
    path_to_md_dir = system.file("module_examples/md/", package = "Magellan")
)
createModuleTemplate("PipelineA", conf.pipeline)

}
\author{
Samuel Wieczorek
}
