% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Process_Template.R
\name{createProcessTemplateForExpert}
\alias{createProcessTemplateForExpert}
\title{Create process template code}
\usage{
createProcessTemplateForExpert(config = NULL)
}
\arguments{
\item{config}{xxx}
}
\description{
xxx
}
\examples{
conf <- list(parent = "pipeline",
name = "process",
steps = c('Description', "Step 1", "Step 2", "Save"),
mandatory = c(TRUE, TRUE, FALSE, TRUE)
)
createProcessTemplateForExpert(conf)

}
\author{
Samuel Wieczorek
}
