% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_funcs.R, R/utils.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{keepDatasets}
\alias{keepDatasets,list-method}
\alias{addDatasets}
\alias{addDatasets,list-method}
\title{Datasets processing}
\usage{
addDatasets(object, dataset, name)

keepDatasets(object = NULL, range = seq(length(object)))
}
\arguments{
\item{object}{An object of class `list`.}

\item{dataset}{`character()` providing the base with respect to which
logarithms are computed. Default is log2.}

\item{name}{A `character()` naming the new array name.}

\item{range}{A interval of integers}
}
\value{
An processed object of the same class as `object`.

An processed object of the same class as `object`.
}
\description{
This manual page describes manipulation methods using [list] objects. In
index or name `i` can be specified to define the array (by name of
index) on which to operate.

The following functions are currently available:

- `keepDatasets(object, range)` keep datasets in object which
are in range

- `addDatasets(object, dataset, name)` add the 'dataset' to the
object (of type list)

- `Save(object, file)` stores the object to a .RData file

This function appends a dataset in the list with customization
if necessary

This function deletes the items not included in the
range parameter

This manual page describes manipulation methods using [list] objects. In
index or name `i` can be specified to define the array (by name of
index) on which to operate.

The following functions are currently available:

- `keepDatasets(object, range)` keep datasets in object which
are in range

- `addDatasets(object, dataset, name)` add the 'dataset' to the
object (of type list)

- `Save(object, file)` stores the object to a .RData file
}
\details{
The object must be of type list. Thetwo functions are implemented here for
of the package which uses MagellanNTK

The object must be of type list. Thetwo functions are implemented here for
of the package which uses MagellanNTK
}
\examples{
data(lldata)
for (i in 1:5){
lldata <- c(lldata, newEL = lldata[[i]])
names(lldata)[i+1] <- paste0('Copy_', i)
}
keepDatasets(lldata, 2:3)


NULL
}
