% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{lldata}
\alias{sub_R25}
\title{lldata dataset}
\format{
An object of class `list()`

An object of class `list()`
}
\usage{
data(lldata)

data(sub_R25)
}
\description{
xxx

xxx
}
\keyword{data}
\keyword{datasets}
