% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_debug_infos.R
\name{Debug_Infos}
\alias{Debug_Infos}
\alias{Debug_Infos_ui}
\alias{Debug_Infos_server}
\title{Module debugging}
\usage{
Debug_Infos_ui(id)

Debug_Infos_server(
  id,
  title = NULL,
  config = reactive({
     NULL
 }),
  rv.dataIn = reactive({
     NULL
 }),
  dataIn = reactive({
     NULL
 }),
  dataOut = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  steps.status = reactive({
     NULL
 }),
  steps.skipped = reactive({
     NULL
 }),
  current.pos = reactive({
     NULL
 }),
  is.enabled = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}

\item{title}{The title of the Panel which contains the debugging tables}

\item{config}{An instance of the class `Config`}

\item{rv.dataIn}{xxx}

\item{dataIn}{A `list()` of data.frames}

\item{dataOut}{A `list()` of data.frames}

\item{steps.enabled}{A `logical()` xxxx}

\item{steps.status}{A `logical()` xxxx}

\item{steps.skipped}{A `logical()` xxxx}

\item{current.pos}{A `integer(1)` which is the indice of the active step.}

\item{is.enabled}{A `logical(1)` xxxx}
}
\value{
NA
}
\description{
A shiny Module which shows the values of the datasets variables
along the different processes of a process nor pipeline.
}
\examples{
#'
if (interactive()) {
    ui <- fluidPage(
        Debug_Infos_ui("tbl")
    )
    server <- function(input, output) {
        Debug_Infos_server(
        id = "tbl", 
        data = reactive({head(iris)
        }))
    }
    shinyApp(ui, server)
}

}
