% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{default_pos}
\alias{default_pos}
\title{xxxx}
\format{
An object of class \code{list} of length 3.
}
\usage{
default_pos
}
\description{
xxx
}
\section{Fields}{

\describe{
\item{\code{btn_style}}{xxxx}
}}

\keyword{datasets}
