% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_variables.R
\name{default_vars}
\alias{default_vars}
\alias{default_funcs}
\alias{default_base_URL}
\alias{default_workflow}
\alias{default_layout}
\alias{default_theme}
\title{Default variables used in MagellanNTK}
\usage{
default_funcs()

default_base_URL()

default_workflow()

default_layout

default_theme(mode)
}
\arguments{
\item{mode}{A `character()` to specifies the running mode of MagellanNTK:
'user' (default) or 'dev'.}
}
\value{
NA
}
\description{
left_panel = left_sidebar + width_sidebar
top_sidebar = top_panel
width_sidebar = width_process_btns
}
\examples{
NULL

}
