% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_btns.R
\name{download_btns}
\alias{download_btns}
\alias{download_btns_ui}
\alias{download_btns_server}
\title{Download_btns shiny module.}
\usage{
download_btns_ui(id, settings = list())

download_btns_server(
  id,
  obj = reactive({
     NULL
 }),
  name,
  colors = reactive({
     NULL
 }),
  tags = reactive({
     NULL
 })
)

download_btns(obj)
}
\arguments{
\item{id}{internal}

\item{settings}{xxx}

\item{obj}{A data.frame}

\item{name}{xxxx.}

\item{colors}{xxx}

\item{tags}{xxx}
}
\value{
NA
}
\description{
A shiny module that shows download buttons in different formats.
}
\examples{
\dontrun{
data(lldata)
shiny::runApp(download_btns(lldata))
}

}
