% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dyn_widgets.R
\name{DynamicWidgets}
\alias{DynamicWidgets}
\alias{dyn_widgets_ui}
\alias{dyn_widgets_server}
\alias{dyn_widgets}
\title{Create dynamic widgets}
\usage{
dyn_widgets_ui(id)

dyn_widgets_server(id)

dyn_widgets()
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module to create a dynamic number of widgets.
}
