% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_error_modal.R
\name{errorModal}
\alias{errorModal}
\alias{errorModal_ui}
\alias{errorModal_server}
\title{Error modal shiny module.}
\usage{
errorModal_ui(id)

errorModal_server(id, msg)

errorModal(msg)
}
\arguments{
\item{id}{internal}

\item{msg}{xxxx}
}
\value{
NA
}
\description{
A shiny module that shows messages in modal.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(errorModal("my error text"))
\dontshow{\}) # examplesIf}
}
