% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mod_Pipeline.R, R/example_mod_Process.R
\name{example_mod_Pipeline}
\alias{example_mod_Pipeline}
\alias{example_mod_Process}
\title{Shiny example module `Pipeline A`}
\usage{
example_mod_Pipeline(verbose = FALSE)

example_mod_Process(verbose = FALSE)
}
\arguments{
\item{verbose}{xxx}
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package Magellan.
This documentation is for developpers who want to create their own pipelines nor processes
to be managed with `Magellan`.

This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package Magellan.
This documentation is for developpers who want to create their own pipelines nor processes
to be managed with `Magellan`.
}
\author{
Samuel Wieczorek

Samuel Wieczorek
}
