% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtraModules.R
\name{extra_module}
\alias{extra_module}
\alias{default_extraModule_code}
\title{## To be customized ##}
\usage{
default_extraModule_code()
}
\arguments{
\item{id}{xxx}
}
\description{
## To be customized ##
}
\examples{
if (interactive()){
data(ft_na)
ui <- foo_ui('query')

server <- function(input, output, session) {
  
  rv <- reactiveValues(
    res = NULL
  )
  ll.tags <- c('None' = 'None', 
  qMetadata.def(typeDataset(ft_na[[1]]))$node)
  
  rv$res <- foo_server('query', 
  reset = reactive({NULL}),
  is.enabled = reactive({NULL})
  )
  
  observeEvent(rv$res$dataOut()$value, ignoreNULL = TRUE, ignoreInit = TRUE, {
    print(rv$res$dataOut()$value)
    print(rv$res$dataOut()$widgets)
  })
}

shinyApp(ui=ui, server=server)
}
}
