% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feat1}
\alias{feat1}
\title{Feature example data}
\format{
An object of class \code{list} of length 4.
}
\usage{
feat1
}
\description{
`feat1` is a small test `Magellan` object for testing and
demonstration. `feat2` is used to demonstrate assay joins. `ft_na`
is a tiny test set that contains missing values used to
demonstrate and test the impact of missing values on data
processing. `se_na2` is an `SummarizedExperiment` with missing
values of mixed origin.
}
\keyword{datasets}
