% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{FindClosingParenthesis}
\alias{FindClosingParenthesis}
\title{xxx}
\usage{
FindClosingParenthesis(text, openPos)
}
\arguments{
\item{text}{xxx}

\item{openPos}{A `integer`}
}
\value{
A `integer`
}
\description{
xxx
}
