% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{G_Settings}
\alias{G_Settings}
\title{#' @title xxx
#' @description xxx
#' @export}
\format{
An object of class \code{list} of length 13.
}
\usage{
G_Settings
}
\description{
#' @title xxx
#' @description xxx
#' @export
}
\keyword{datasets}
