% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{Generate_code_for_ValidationBtns_renderUI}
\alias{Generate_code_for_ValidationBtns_renderUI}
\title{Code for validation buttons renderUI()}
\usage{
Generate_code_for_ValidationBtns_renderUI(steps)
}
\arguments{
\item{steps}{xxx}
}
\description{
This function Generates dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
steps <- paste0('step', 1:3)
code <- Generate_code_for_ValidationBtns_renderUI(steps)
cat(code)
}

}
\author{
Samuel Wieczorek
}
