% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{Generate_RenderUI_Code_For_Single_Widgets}
\alias{Generate_RenderUI_Code_For_Single_Widgets}
\title{Code for declaring xxx}
\usage{
Generate_RenderUI_Code_For_Single_Widgets(widgets)
}
\arguments{
\item{widgets}{xxx}
}
\description{
This function Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
code <- Generate_RenderUI_Code_For_Single_Widgets(widgets)
cat(code)
}

}
\author{
Samuel Wieczorek
}
