% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{generic_mod_open_dataset}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{Change the default functions in `MagellanNTK`}
\usage{
open_dataset_ui(id)

open_dataset_server(
  id,
  class = NULL,
  extension = NULL,
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

open_dataset(extension = NULL)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{class}{The type of file to allow for other packages dataset}

\item{extension}{The extension file allowed}

\item{remoteReset}{An `integer` which acts as a remote command to reset the
module. Its value is incremented on a external event and it is used to
trigger an event in this module}

\item{is.enabled}{A `boolean`. This variable is used as a remote command to
specify if the corresponding module is enabled/disabled in the calling module
of upper level.
For example, if this module is disabled, then this variable is set to TRUE.
Then, all the widgets will be disabled. If not, the enabling/disabling of
widgets is deciding by this module.}
}
\value{
A Shiny app
}
\description{
This module allows to change
}
\examples{
if (interactive()) {
  open_dataset(extension = "rdata")
}

}
