% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_AdditionalModule_Core_Code}
\alias{Get_AdditionalModule_Core_Code}
\title{Code for declaring xxx}
\usage{
Get_AdditionalModule_Core_Code(w.names, rv.custom.names = NULL)
}
\arguments{
\item{w.names}{xxx}

\item{rv.custom.names}{xxx}
}
\value{
NA
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
widgets <- c('widget1', 'widget2')
Get_AdditionalModule_Core_Code(widgets)

}
\author{
Samuel Wieczorek
}
