% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_Declare_rv_custom}
\alias{Get_Code_Declare_rv_custom}
\title{Code for declaring rv.custom.default.values reactive variable}
\usage{
Get_Code_Declare_rv_custom(rv.custom.names = NULL)
}
\arguments{
\item{widgets.names}{A `list` containing the names of the widgets in all
steps of the module.}
}
\description{
This function create the source code needed inside a module
to declare the reactive variable called 'rv.custom.default.values'.
# Declaration of the variables that will contain the values of the user
variables.
To avoid confusion, the first string is the name of the step while the second is the name
of the widget
}
\examples{
\dontrun{
custom <- list(foo1 = list(), foo2 = 3)
code <- Get_Code_Declare_rv_custom(names(custom))
cat(code)
}

}
\author{
Samuel Wieczorek
}
