% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_Declare_widgets}
\alias{Get_Code_Declare_widgets}
\title{R code to update the 'config' variable of a pipeline module}
\usage{
Get_Code_Declare_widgets(widgets.names = NULL)
}
\arguments{
\item{widgets.names}{A `list` containing the names of the widgets in all
steps of the module.}

\item{id}{The 'id' of the module that calls this function.}
}
\description{
This function generates the necessary code to
modify the variable 'config' (slots steps and mandatory). It adds
a 'Description' step and a TRUE value at the beginning of the 'steps' and 'mandatory'
 list, erases all white spaces for the names of the steps.

This function gives the necessary code to
provide the module (ui() and server() functions) relative to
the Description of a pipeline.
The work differs from the process strategy because a 
pipeline does not have its own UIs as it is only a bridge
between the user shiny app and the UIs of the processes.

This function gives the necessary code to
provide the module (ui() and server() functions) relative to
the Description of a pipeline.
The work differs from the process strategy because a 
pipeline does not have its own UIs as it is only a bridge
between the user shiny app and the UIs of the processes.

This function create the source code needed inside a module
to declare the reactive variable called 'widgets.default.values'.
# Declaration of the variables that will contain the values of the widgets
To avoid confusion, the first string is the name of the step while the second is the name
of the widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
code <- Get_Code_Declare_widgetsDefaultValues(widgets)
cat(code)
}

}
\author{
Samuel Wieczorek
}
