% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{Get_Code_Declare_widgetsDefaultValues}
\alias{Get_Code_Declare_widgetsDefaultValues}
\title{Code for declaring widgets.default.values reactive variable}
\usage{
Get_Code_Declare_widgetsDefaultValues(widgets.names)
}
\arguments{
\item{widgets.names}{xxx}
}
\description{
This function create the source code needed inside a module
to declare the reactive variable called 'widgets.default.values'.
# Declaration of the variables that will contain the values of the widgets
To avoid confusion, the first string is the name of the step while the second is the name
of the widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
code <- Get_Code_Declare_widgetsDefaultValues(widgets)
cat(code)
}

}
\author{
Samuel Wieczorek
}
