% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_AddMod_observeEvents}
\alias{Get_Code_for_AddMod_observeEvents}
\title{Code for declaring xxx}
\usage{
Get_Code_for_AddMod_observeEvents()
}
\value{
NA
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
Get_Code_for_AddMod_observeEvents()

}
\author{
Samuel Wieczorek
}
