% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_dataOut}
\alias{Get_Code_for_dataOut}
\title{Code for xxx}
\usage{
Get_Code_for_dataOut()
}
\description{
Returned value of the process
}
\examples{
\dontrun{
code <- Get_Code_for_dataOut()
cat(code)
}

}
\author{
Samuel Wieczorek
}
