% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_Description_renderUI}
\alias{Get_Code_for_Description_renderUI}
\title{Generate code for a pipeline Description module}
\usage{
Get_Code_for_Description_renderUI(id)
}
\arguments{
\item{id}{The 'id' of the module that calls this function.}
}
\description{
This function gives the necessary code to
provide the module (ui() and server() functions) relative to
the Description of a pipeline.
The work differs from the process strategy because a 
pipeline does not have its own UIs as it is only a bridge
between the user shiny app and the UIs of the processes.
}
