% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_General_observeEvents}
\alias{Get_Code_for_General_observeEvents}
\title{Code for declaring xxx}
\usage{
Get_Code_for_General_observeEvents()
}
\value{
NA
}
\description{
xxx
}
\examples{
Get_Code_for_General_observeEvents()

}
\author{
Samuel Wieczorek
}
