% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_observeEvent_remoteReset}
\alias{Get_Code_for_observeEvent_remoteReset}
\title{Code for xxxx}
\usage{
Get_Code_for_observeEvent_remoteReset()
}
\description{
xxxx
}
\examples{
\dontrun{
code <- Get_Code_for_observeEvent_remoteReset()
cat(code)
}

}
\author{
Samuel Wieczorek
}
