% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_observeEvent_stepsEnabled}
\alias{Get_Code_for_observeEvent_stepsEnabled}
\title{Code for declaring xxx}
\usage{
Get_Code_for_observeEvent_stepsEnabled()
}
\description{
xxx
}
\examples{
\dontrun{
code <- Get_Code_for_observeEvent_stepsEnabled()
cat(code)
}

}
\author{
Samuel Wieczorek
}
