% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_ObserveEvent_widgets}
\alias{Get_Code_for_ObserveEvent_widgets}
\title{Code for declaring xxx}
\usage{
Get_Code_for_ObserveEvent_widgets(widgets.names = NULL)
}
\arguments{
\item{widgets.names}{A `list` containing the names of the widgets in all
steps of the module.}
}
\value{
NA
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
widgets <- paste0("widget", 1:3)
Get_Code_for_ObserveEvent_widgets(widgets)

}
\author{
Samuel Wieczorek
}
