% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_for_rv_reactiveValues}
\alias{Get_Code_for_rv_reactiveValues}
\title{Code for declaring widgets.default.values reactive variable}
\usage{
Get_Code_for_rv_reactiveValues()
}
\value{
A `string` containing some R code
}
\description{
This function createxxx
}
\examples{
Get_Code_for_rv_reactiveValues()

}
\author{
Samuel Wieczorek
}
