% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Worflow_Core_Code}
\alias{Get_Worflow_Core_Code}
\title{Code for declaring xxx}
\usage{
Get_Worflow_Core_Code(w.names, rv.custom.names)
}
\arguments{
\item{w.names}{xxx}

\item{rv.custom.names}{xxx}
}
\value{
NA
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
widgets <- c('widget1', 'widget2')
custom <- list(foo1 = list(), foo2 = 3)
Get_Worflow_Core_Code(widgets, names(custom))

}
\author{
Samuel Wieczorek
}
