% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetCode_observeEvent_isSkipped}
\alias{GetCode_observeEvent_isSkipped}
\title{Skipping an entire process}
\usage{
GetCode_observeEvent_isSkipped()
}
\description{
The parameter is.skipped() is set by the caller and tells the process
if it is skipped or not (remote action from the caller)
}
