% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{GetCode_ResetChildren}
\alias{GetCode_ResetChildren}
\title{xxx}
\usage{
GetCode_ResetChildren()
}
\value{
Source code for function ResetChildren
}
\description{
et to skipped all steps of the current object
}
\author{
Samuel Wieczorek
}
