% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{GetConfig}
\alias{GetConfig}
\title{Get config variable from workflow source code}
\usage{
GetConfig(s)
}
\arguments{
\item{s}{The complete source code of the process module}
}
\value{
A list containing the configuration of the module
}
\description{
Analyze the source code to extract the config variable
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
