% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{GetExtension}
\alias{GetExtension}
\title{Get file extension}
\usage{
GetExtension(name)
}
\arguments{
\item{name}{A complete filename}
}
\value{
The extension of the given filename
}
\description{
Get file extension
}
\examples{
GetExtension('foo.xlsx')
}
