% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetFirstMandatoryNotValidated}
\alias{GetFirstMandatoryNotValidated}
\title{Get the first mandatory step not validated}
\usage{
GetFirstMandatoryNotValidated(range, rv)
}
\arguments{
\item{range}{xxx}

\item{rv}{A `list` with at least two slots :
* mandatory: xxx
* steps.status: xxx}
}
\value{
An integer which is the indice of the identified step in the vector
rv$steps.status
}
\description{
Get the first mandatory step not validated
}
\examples{
NULL

}
