% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetHistory}
\alias{GetHistory}
\title{Get the last validated step before current position.}
\usage{
GetHistory(dataIn, x)
}
\arguments{
\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{x}{A `character` which is the name of an assay in the dataIn}
}
\value{
A `data.frame()`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL

}
