% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pipeline.R
\name{GetMaxValidated_BeforeCurrentPos}
\alias{GetMaxValidated_BeforeCurrentPos}
\title{xxx}
\usage{
GetMaxValidated_BeforeCurrentPos()
}
\value{
Nothing
}
\description{
This function calls the server part of each module composing the pipeline
}
