% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{Get the last validated step before current position.}
\usage{
GetMaxValidated_BeforePos(pos = NULL, rv)
}
\arguments{
\item{pos}{A `integer()` which is the indice of the active position.}

\item{rv}{A `list()` which stores reactiveValues()}
}
\value{
A `integer()`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
pos <- 3
rv <- list(steps.status = c(1,1,0,1,0,0), current.pos = 3)
GetMaxValidated_BeforePos(pos, rv)

}
