% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{GetPackageVersion}
\alias{GetPackageVersion}
\title{Get the version of a package installed in the local R distribution}
\usage{
GetPackageVersion(pkg)
}
\arguments{
\item{pkg}{The name of a package}
}
\value{
A `character()`
}
\description{
Get the version of a package installed in the local R distribution
}
\examples{
GetPackageVersion('MagellanNTK')
}
