% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{globalvariables}
\alias{globalvariables}
\alias{tl_v_next_icon}
\alias{stepStatus}
\alias{default_pos}
\alias{listBrewerPalettes}
\title{Global variables}
\usage{
tl_v_next_icon

stepStatus

default_pos

listBrewerPalettes
}
\value{
A collection of exported constants, lists and UI configuration objects.
}
\description{
Global constants and configuration variables used throughout the package.
}
