% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{initComplete}
\alias{initComplete}
\title{Initialisation function for DT}
\usage{
initComplete()
}
\value{
literal JavaScript code
}
\description{
Initialisation function for DT
}
\examples{
initComplete()
}
