% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{InitializeHistory}
\alias{InitializeHistory}
\title{Get the last validated step before current position.}
\usage{
InitializeHistory()
}
\value{
A `data.frame()` with four columns: 'Process', 'Step', 'Parameter'
and 'Value'
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
InitializeHistory()

}
