% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Insert_Call_to_Config}
\alias{Insert_Call_to_Config}
\title{Insert source code for config()}
\usage{
Insert_Call_to_Config(name)
}
\arguments{
\item{name}{xxx}
}
\value{
A `character()` containing R source code
}
\description{
Insert source code for config()
}
