% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertCode}
\alias{insertCode}
\title{xxx}
\usage{
insertCode(f)
}
\arguments{
\item{f}{a filename}
}
\value{
A `character(1)` vector containing as a character the
source code of the file 'f'. Ready to use with eval(str2expression())
}
\description{
Returns the date and time in timestamp UNIX format.
}
\author{
Samuel Wieczorek
}
