% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Magellan.compliant}
\alias{is.Magellan.compliant}
\title{Checks if the object is compliant with MagellanNTK}
\usage{
is.Magellan.compliant(obj)
}
\arguments{
\item{obj}{xxx}
}
\description{
Checks and accept the following data formats:
* An instance of class `MSnSet`
* An instance of class `MultiAssayExperiment`
* An instance of class `SummarizedExperiment`
* An instance of class `data.frame`
* An instance of class `matrix`
* A list of instances of class `MSnSet`
* A list of instances of class `SummarizedExperiment`
* A list of instances of class `data.frame`
* A list of instances of class `matrix`
}
\examples{
is.Magellan.compliant(data.frame())


ll <- list(data.frame(), data.frame())
is.Magellan.compliant(ll)

}
