% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isSubstr}
\alias{isSubstr}
\title{Substring test}
\usage{
isSubstr(pattern, target)
}
\arguments{
\item{pattern}{A `character()`}

\item{target}{The `character()` to look for into the pattern}
}
\value{
A boolean
}
\description{
Substring test
}
\examples{
isSubstr("cde", "abcdefghi")

}
