% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{xxxx}
\usage{
launch_app(mode = "user")
}
\arguments{
\item{mode}{A  `character(1)` which indicates xxxx. Available values are: 
'user' (default) and 'dev'.}
}
\description{
xxxx
}
