% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{lldata}
\alias{lldata}
\title{lldata dataset}
\format{
An object of class `MultiAssayExperiment`
}
\usage{
data(lldata)
}
\description{
An instance of the class `MultiAssayExperiment` which contains
one assay (instance of class `SummarizedExperiment`):
'Convert': Contains a data.frame with 6 columns and 100 rows fullfilled with
'0'.
It also contains a slot 'history' in the metadata().
This dataset is the most
bacis one to run the example pipeline and learn how to work with
MagellanNTK
}
\keyword{data}
\keyword{datasets}
