% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{lldata12}
\alias{lldata12}
\title{lldata12 dataset}
\format{
An object of class `MultiAssayExperiment`
}
\usage{
data(lldata12)
}
\description{
An instance of the class `MultiAssayExperiment` which contains
two assays (instance of class `SummarizedExperiment`):
'Convert': Contains a data.frame with 6 columns and 100 rows fullfilled with
'0'.
'DataGeneration' which stores the result of the 'DataGeneration' step of the
demo pipeline
'Preprocessing' which stores the result of the 'DataGeneration' and
'Preprocessing' steps of the demo pipeline
Each assay contains a slot 'history' in the metadata().
}
\keyword{data}
\keyword{datasets}
