% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_workflow.R
\name{Load_Dataset}
\alias{Load_Dataset}
\alias{Load_Workflow_ui}
\alias{Load_Workflow_server}
\title{Load dataset shiny module}
\usage{
Load_Workflow_ui(id)

Load_Workflow_server(
  id,
  path = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
\examples{

#if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    mod_openfile_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    mod_openfile_server("demo", path = reactive({NULL}))
  }
  
  shinyApp(ui, server)
#}
}
