% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_workflow.R
\name{load_workflow}
\alias{load_workflow}
\alias{mod_load_workflow_ui}
\alias{mod_load_workflow_server}
\title{Load dataset shiny module}
\usage{
mod_load_workflow_ui(id)

mod_load_workflow_server(
  id,
  path = reactive({
     NULL
 }),
  usermod = "user"
)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{path}{xxx}

\item{usermod}{xxx}
}
\value{
NA
}
\description{
A shiny Module to load a dataset.
}
