% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_code.R
\name{LoadCode}
\alias{LoadCode}
\title{Load workflows functions}
\usage{
LoadCode(name, path, recursive = FALSE)
}
\arguments{
\item{name}{A `character()` to indicate the name of the workflow. It can be
either a single string (which represents a workflow without a parent) or two
strings separated by '_' (in that case, it is a workflow with a parent 
workflow).}

\item{path}{A `character()` to indicate the directory where to find the source
files of all workflows. This directory must also contains a 'md' directory
that contains the md files corresponding to the workflows.}

\item{recursive}{xxx}
}
\description{
This function xx
}
\examples{
.path <- system.file('extdata/module_examples', package='MagellanNTK')

.name <- 'PipelineA_Process1'
LoadCode(name = .name, path = .path)

.path <- system.file('extdata/module_examples', package='MagellanNTK')
.name <- 'PipelineA'
LoadCode(name = .name, path = .path)
}
