% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{LocalReset}
\alias{LocalReset}
\title{xxx}
\usage{
LocalReset(steps, nav.mode)
}
\arguments{
\item{steps}{xxx}

\item{nav.mode}{xxx}
}
\value{
A list of four items:
* dataIn:  xx
* dataOut: xxx
* current.pos: xxx
* steps.status: xxx
}
\description{
xxx
}
