% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Launch_Magellan.R
\name{Magellan}
\alias{Magellan}
\title{Pipeline launch class}
\description{
This class is a workaround to make work shinyjs::disable/enable functions
within renderUI:
https://github.com/daattali/shinyjs/issues/74
https://github.com/daattali/shinyjs/issues/25
Its goal is to activate automatically the pipeline when a 
dataset is loaded
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Magellan$new()}}
\item \href{#method-Launch_Pipeline}{\code{Magellan$Launch_Pipeline()}}
\item \href{#method-ui}{\code{Magellan$ui()}}
\item \href{#method-server}{\code{Magellan$server()}}
\item \href{#method-clone}{\code{Magellan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magellan$new(id, name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Launch_Pipeline"></a>}}
\if{latex}{\out{\hypertarget{method-Launch_Pipeline}{}}}
\subsection{Method \code{Launch_Pipeline()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magellan$Launch_Pipeline(pipe.name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magellan$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magellan$server(dataIn = reactive({
    NULL
}))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Magellan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
