% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanNTK.R
\name{magellanNTK}
\alias{magellanNTK}
\alias{MagellanNTK_ui}
\alias{MagellanNTK_server}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK_ui(id, sidebarSize = "medium")

MagellanNTK_server(
  id,
  workflow.path = reactive({
     NULL
 }),
  workflow.name = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "user"
)

MagellanNTK(
  workflow.path = NULL,
  workflow.name = NULL,
  verbose = FALSE,
  usermod = "user",
  sidebarSize = "medium",
  ...
)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{sidebarSize}{The width of the sidebar. Available values are 'small', 'medium', 'large'}

\item{workflow.path}{A `character()` which is the path to the directory which 
contains the files and directories of the pipeline.}

\item{workflow.name}{A `character()` which is the name of the pipeline to
launch and run whithin the framework of MagellanNTK. It designs the name of 
a directory which contains the files and directories of the pipeline.}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}

\item{usermod}{A `character()` to specifies the running mode of MagellanNTK: 
'user' (default) or 'dev'. For more details, please refer to the document 
'Inside MagellanNTK'}

\item{...}{Additional parameters}
}
\value{
A shiny app
}
\description{
Main Shiny application
}
\examples{
if (interactive()) {
    MagellanNTK()
    }

}
