% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MagellanNTK2}
\alias{MagellanNTK2}
\title{Main Shiny application}
\usage{
MagellanNTK2(funcs = default.funcs(), verbose = FALSE)
}
\arguments{
\item{funcs}{An list containing th following items:
* title: xxx
* base_URL: xxx}

\item{verbose}{A `boolean(1)`}
}
\description{
xxx
}
\details{
The list of customizable funcs (param `funcs`) contains the following items:

 These are the default values where each item points to a default fucntion
 implemented into MagellanNTK.
 
 The user can modify these values by two means:
 * setting the values in the parameter to pass to the function 
 `MagellanNTK()`,
 * inside the UI of MagellanNTK, in the settings panels
}
\examples{
\dontrun{

# launch without initial config
shiny::runApp(MagellanNTK())

# launch with config
funcs <- list(convert_dataset = "DaparToolshed::convert_dataset",
open_dataset = "DaparToolshed::open_dataset",
open_demoDataset = "DaparToolshed::open_demoDataset",
view_dataset = "omXplore::view_dataset",
download_dataset = "MagellanNTK::download_dataset",
export_dataset = "MagellanNTK::export_dataset",
infos_dataset = "DaparToolshed::infos_dataset",
addDatasets = "DaparToolshed::addDatasets",
keepDatasets = "DaparToolshed::keepDatasets" )

funcs <- list(convert_dataset = "MagellanNTK::convert",
  open_dataset = "MagellanNTK::open_dataset",
  open_demoDataset = "MagellanNTK::open_demoDataset",
  view_dataset = "MagellanNTK::view_dataset",
  export_dataset = "MagellanNTK::export_dataset",
  infos_dataset = "MagellanNTK::infos_dataset",
  download_dataset = "MagellanNTK::download_dataset",
  addDatasets = "MagellanNTK::addDatasets",
  keepDatasets = "MagellanNTK::keepDatasets")

MagellanNTK(funcs)
}

}
