% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanNTK.R
\name{MagellanNTK_ui}
\alias{MagellanNTK_ui}
\title{The application User-Interface}
\usage{
MagellanNTK_ui(id)
}
\arguments{
\item{request}{Internal parameter for `{shiny}`. 
DO NOT REMOVE.}
}
\description{
The application User-Interface
}
