% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dl.R
\name{mod-dl}
\alias{mod-dl}
\alias{mod_dl_ui}
\alias{mod_dl_server}
\title{mod_dl}
\usage{
mod_dl_ui(id)

mod_dl_server(
  id,
  dataIn = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    data(ft_na)
    ui <- mod_dl_ui("dl")

    server <- function(input, output, session) {
        mod_dl_server(
            id = "dl",
            dataIn = reactive({ft_na})
        )}
    shinyApp(ui = ui, server = server)
}

}
