% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bsModal.R
\name{mod_bsmodal}
\alias{mod_bsmodal}
\alias{mod_bsmodal_ui}
\alias{mod_bsmodal_server}
\title{Predefined modal}
\usage{
mod_bsmodal_ui(id)

mod_bsmodal_server(
  id,
  label = "Edit md",
  title = NULL,
  width = NULL,
  uiContent = NULL,
  shiny.module = NULL
)

mod_bsmodal(
  title = "test",
  ui.func = NULL,
  ui.params = list(),
  server.func = NULL,
  server.params = list()
)
}
\arguments{
\item{id}{A `character(1)` which is the id of the instance of the module}

\item{label}{xxx}

\item{title}{A `character(1)`}

\item{width}{A `character(1)` indicating the size of the modal window. Can 
be "s" for small (the default), "m" for medium, or "l" for large.}

\item{uiContent}{The content of the modal dialog.}

\item{shiny.module}{xxx}
}
\value{
A Shiny modal-dialog
}
\description{
Displays of formatted modal-dialog with 'Cancel' and 
'Ok' buttons.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_bsmodal())
\dontshow{\}) # examplesIf}
}
