% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_check_updates.R
\name{mod_check_updates}
\alias{mod_check_updates}
\alias{mod_check_updates_ui}
\alias{mod_check_updates_server}
\title{mod_check_updates_ui and mod_check_updates_server}
\usage{
mod_check_updates_ui(id)

mod_check_updates_server(id)

mod_check_updates()
}
\arguments{
\item{id}{shiny id}
}
\value{
xxxx
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_check_updates())
\dontshow{\}) # examplesIf}
}
