% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_choose_pipeline.R
\name{mod_choose_pipeline}
\alias{mod_choose_pipeline}
\alias{choose_pipeline_ui}
\alias{choose_pipeline_server}
\title{choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
choose_pipeline_ui(id)

choose_pipeline_server(id, dataType = NULL, package = NULL)
}
\arguments{
\item{id}{shiny id}

\item{dataType}{xxx}

\item{package}{xxx}
}
\description{
A shiny Module.
}
\keyword{internal}
