% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert_dataset.R
\name{mod_convert_dataset}
\alias{mod_convert_dataset}
\alias{convert_dataset_ui}
\alias{convert_dataset_server}
\alias{convert_dataset}
\title{convert_dataset_ui and convert_dataset_server}
\usage{
convert_dataset_ui(id)

convert_dataset_server(id)

convert_dataset()
}
\arguments{
\item{id}{xxx}
}
\value{
A list
}
\description{
A shiny Module.
}
\examples{
\dontrun{
shiny::runApp(convert_dataset())
}

}
