% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dl.R
\name{mod_dl}
\alias{mod_dl}
\alias{mod_dl_ui}
\alias{mod_dl_server}
\title{mod_dl}
\usage{
mod_dl_ui(id)

mod_dl_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = reactive({
     "csv"
 }),
  widget.type = reactive({
     "Link"
 }),
  name = reactive({
     "foo"
 }),
  excel.style = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are `csv` (default), `RData` and `Excel`.}

\item{widget.type}{Available values are `Button` and `Link` (default).}

\item{name}{internal}

\item{excel.style}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{

if (interactive()){
  library(shinyjqui)
  library(DT)
  library(shinyjs)
  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  data(data_na)
  
  ui <- fluidPage(
    mod_download_btns_ui("dl")
  )
  
  server <- function(input, output, session) {
    
    mod_download_btns_server(
      id = "dl",
      df.data = reactive({data_na$array1}),
      name = reactive({'foo'}),
      extension = reactive({c('Excel', 'csv', 'RData')}),
      widget.type = reactive({c('Link', 'Button', 'Link')})
    )
  }
  
  shinyApp(ui, server)
}
}
