% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_dataset.R
\name{mod_export_dataset}
\alias{mod_export_dataset}
\alias{export_dataset_ui}
\alias{export_dataset_server}
\alias{export_dataset}
\title{Export dataset Shiyny app}
\usage{
export_dataset_ui(id)

export_dataset_server(id, dataIn)

export_dataset(data)
}
\arguments{
\item{id}{xxx}
}
\value{
A list
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(lldata)
shiny::runApp(export_dataset(lldata))
\dontshow{\}) # examplesIf}
}
