% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT_ui}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(
  id,
  table2show,
  withBtns = NULL,
  showRownames = FALSE,
  dom = "Bt",
  style = reactive({     NULL })
)
}
\arguments{
\item{id}{xxx}

\item{table2show}{xxx}

\item{withBtns}{xxx}

\item{showRownames}{xxxx}

\item{dom}{xxx}

\item{style}{xxx}
}
\value{
xxx
}
\description{
A shiny Module.

xxx
}
\examples{
\dontrun{

ui <- fluidPage(
  mod_format_DT_ui('tbl')
)
server <- function(input, output){
  mod_format_DT_server(id = 'tbl',table2show = reactive({head(iris)}))
}
shinyApp(ui, server)
}

}
