% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_homepage.R
\name{mod_homepage}
\alias{mod_homepage}
\alias{mod_homepage_ui}
\alias{mod_homepage_server}
\title{mod_homepage_ui and mod_homepage_server}
\usage{
mod_homepage_ui(id)

mod_homepage_server(
  id,
  mdfile = file.path(system.file("www/md", package = "MagellanNTK"), "Presentation.Rmd"),
  dataset = reactive({
     NULL
 })
)

mod_homepage()
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{mdfile}{The path to the Rmd file which describes the pipeline}

\item{dataset}{An instance of the class `MultiAssayExperiment`}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
  mod_homepage()
}

}
