% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_launch_magellan.R
\name{mod_launch_magellan}
\alias{mod_launch_magellan}
\alias{mod_launch_magellan_ui}
\alias{mod_launch_magellan_server}
\title{launch_magellan UI Function}
\usage{
mod_launch_magellan_ui(id)

mod_launch_magellan_server(id)

mod_launch_magellan()
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(mod_launch_magellan())
}

}
