% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_package.R
\name{mod_load_package}
\alias{mod_load_package}
\alias{mod_load_package_ui}
\alias{mod_load_package_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
mod_load_package_ui(id)

mod_load_package_server(id, funcs = NULL)
}
\arguments{
\item{id}{shiny id}

\item{pkg}{xxx}
}
\description{
This module is used to load the package which contains all the code
to work with prostar.2.0.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
funcs <- list(convert_dataset = "DaparToolshed::convert_dataset",
open_dataset = "MagellanNTK::open_dataset",
open_demoDataset = "MagellanNTK::open_demoDataset",
view_dataset = "omXplore::view_dataset",
infos_dataset = "MagellanNTK::infos_dataset",
addDatasets = "MagellanNTK::addDatasets",
keepDatasets = "MagellanNTK::keepDatasets"
)
shiny::runApp(load_package(funcs))

\dontshow{\}) # examplesIf}
}
\author{
Samuel Wieczorek
}
