% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_loading_page.R
\name{mod_loading_page}
\alias{mod_loading_page}
\alias{mod_loading_page_ui}
\alias{mod_loading_page_server}
\title{mod_loading_page_ui and mod_loading_page_server}
\usage{
mod_loading_page_ui(id)

mod_loading_page_server(id)
}
\arguments{
\item{id}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\examples{
if(interactive(){
library(shinyjs)
library(shiny)
ui <- mod_loading_page_ui("mod_info")
server <- function(input, output, session) {
  mod_loading_page_server("mod_info")
}

shinyApp(ui, server)
}
}
\keyword{internal}
