% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainapp.R
\name{mod_main_page}
\alias{mod_main_page}
\alias{mainapp_ui}
\alias{mainapp_server}
\alias{mainapp}
\title{mod_main_page_ui and mod_loading_page_server}
\usage{
mainapp_ui(id, session, size = "300px")

mainapp_server(
  id,
  workflow.name = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "user"
)

mainapp(usermod = "user")
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{session}{shiny internal}

\item{size}{The width of the sidebar.in pixels}

\item{workflow.name}{A `character()` which is the name of the pipeline to
launch and run whithin the framework of MagellanNTK. It designs the name of
a directory which contains the files and directories of the pipeline.}

\item{workflow.path}{A `character()` which is the path to the directory which
contains the files and directories of the pipeline.}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON
(TRUE) the verbose mode for logs.}

\item{usermod}{A `character()` to specifies the running mode of MagellanNTK:
'user' (default) or 'dev'.}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
  mainapp()
}

}
