% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R
\name{mod_nav_pipeline_ui}
\alias{mod_nav_pipeline_ui}
\alias{mod_nav_pipeline_server}
\title{xxx}
\usage{
mod_nav_pipeline_ui(id)

mod_nav_pipeline_server(
  id,
  dataIn = reactive({     NULL }),
  is.enabled = reactive({     TRUE }),
  remoteReset = reactive({     FALSE }),
  is.skipped = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{is.enabled}{xxx}

\item{remoteReset}{xxx}

\item{is.skipped}{xxx}
}
\value{
xxx
}
\description{
xxxxxx

xxxxxx

xxx
}
\section{Fields}{

\describe{
\item{\code{modal_txt}}{xxx}
}}

\examples{
\donttest{
library(shiny)
library(shinyBS)
ui <- fluidPage(
  mod_nav_pipeline_ui('Protein')
)
server <- function(input, output){
mod_nav_pipeline_server(id = 'Protein',
                          dataIn = reactive({QFeatures::feat1})
  )
}
shinyApp(ui, server)
}

}
