% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R
\name{mod_nav_pipeline_server}
\alias{mod_nav_pipeline_server}
\title{xxx}
\usage{
mod_nav_pipeline_server(
  id,
  dataIn = reactive({     NULL }),
  is.enabled = reactive({     TRUE }),
  remoteReset = reactive({     FALSE }),
  is.skipped = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{is.enabled}{xxx}

\item{remoteReset}{xxx}

\item{is.skipped}{xxx}

\item{name}{The name of the process to update}
}
\value{
xxx
}
\description{
xxxxxx

xxx
}
\section{Fields}{

\describe{
\item{\code{modal_txt}}{xxx}
}}

\examples{
library(shiny)
library(shinyBS)
library(crayon)
library(MSPipelines)
ui <- fluidPage(
  mod_nav_pipeline_ui('Protein')
)
server <- function(input, output){
  utils::data(Exp1_R25_prot, package='DAPARdata2')
mod_nav_pipeline_server(id = 'Protein',
                          dataIn = reactive({Exp1_R25_prot})
  )
}
shinyApp(ui, server)

}
