% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_process.R
\name{mod_nav_process_ui}
\alias{mod_nav_process_ui}
\title{xxx}
\usage{
mod_nav_process_ui(id)
}
\arguments{
\item{id}{xxx}
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package Magellan
}
