% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_change_assay.R, R/mod_open_dataset.R,
%   R/mod_popover_for_help.R, R/mod_settings.R
\name{mod_change_assay}
\alias{mod_change_assay}
\alias{mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\alias{popover_for_help}
\alias{mod_settings}
\title{mod_open_dataset_ui and mod_open_dataset_server}
\usage{
mod_change_assay(ll.se, indice)

open_dataset_ui(id)

open_dataset_server(id)

open_dataset()

popover_for_help(title, content)

mod_settings(obj)
}
\arguments{
\item{id}{xxx}
}
\value{
A list
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
shiny::runApp(open_dataset())
}

}
\keyword{internal}
