% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sweetAlert.R
\name{mod_sweetAlert}
\alias{mod_sweetAlert}
\alias{mod_SweetAlert_ui}
\alias{mod_SweetAlert_server}
\alias{mod_SweetAlert}
\title{A shiny app to display a modal}
\usage{
mod_SweetAlert_ui(id)

mod_SweetAlert_server(
  id,
  title = NULL,
  text = NULL,
  showClipBtn = TRUE,
  type = "error"
)

mod_SweetAlert(title, text, type = "warning")
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{title}{The title to be displayed in the window}

\item{text}{The titmessage to be displayed in the window}

\item{showClipBtn}{A `boolean` to indicates if the copy to clipboard button
is shown or not.}

\item{type}{A `character()` for the type of message: 'error', 'warning', etc.}
}
\value{
A shiny App
}
\description{
This module is not directly used by MagellanNTK core functions.
It is rather a useful tool for third part pipelines and processes.
}
\examples{
if (interactive()) {
  mod_SweetAlert("my title", "my message")
}

}
