% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timeline_v.R
\name{mod_timeline_v}
\alias{mod_timeline_v}
\alias{mod_timeline_v_ui}
\alias{mod_timeline_v_server}
\title{Vertical timeline module}
\usage{
mod_timeline_v_ui(id)

mod_timeline_v_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{The `id` of the server}

\item{config}{A static `list` (not `reactiveList`) containing the same 
elements as the process module.}

\item{status}{A `reactive vector` which contain the status (validated,
skipped or undone) of each step of the process module. Its length is equal
to the number of steps.}

\item{position}{A `reactive integer` that reflects the position of the
current (active) step.}

\item{enabled}{A `reactive vector` of length the number of steps and which
indicate whether the step is enabled or disabled.}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if(interactive()){
  library(shiny)
  library(shinyWidgets)
  library(shinyjs)

  options(shiny.fullstacktrace = TRUE)
  

ui <- fluidPage(
  actionButton("prevpos", GlobalSettings$tl_v_prev_icon),
  actionButton("nextpos", GlobalSettings$tl_v_next_icon),
  mod_timeline_v_ui('TLv')
)



server <- function(input, output){
  
  rv <- reactiveValues(
    direction = 0,
    status = c(0, 1, 0, 0),
    current.pos = 1,
    tl.tags.enabled = c(1, 1, 1, 1),
    position = NULL
  )
  
  
  config <- Config(
    name = 'Process1',
    mode = 'process',
    steps = c('Description', 'Step 1', 'Step 2', 'Save'),
    mandatory = c(TRUE, FALSE, TRUE, TRUE),
    path_to_md_dir = system.file('extdata/module_examples/md/', package='MagellanNTK')
  )
  
  
  
  mod_timeline_v_server(id = 'TLv',
    config = config,
    status = reactive({rv$status}),
    position = reactive({rv$current.pos}),
    enabled = reactive({rv$tl.tags.enabled})
  )
  
  observeEvent(input$nextpos,{
    if (rv$current.pos != length(config@steps))
      rv$current.pos <- rv$current.pos + 1
  })
  
  observeEvent(input$prevpos,{
    if (rv$current.pos != 1)
      rv$current.pos <- rv$current.pos - 1
  })
}


shinyApp(ui, server)
}
}
\author{
Samuel Wieczorek
}
