% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timeline_v.R
\name{mod_timeline_v_server}
\alias{mod_timeline_v_server}
\title{timeline Server Function}
\usage{
mod_timeline_v_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{xxx}

\item{config}{xxx}

\item{status}{xxx}

\item{position}{xxx}

\item{enabled}{xxx}
}
\description{
timeline Server Function
}
