% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_modalDialog.R
\name{modalDialog}
\alias{modalDialog}
\alias{modalDialog_ui}
\alias{modalDialog_server}
\title{Predefined modal}
\usage{
modalDialog_ui(id)

modalDialog_server(
  id,
  title = NULL,
  width = NULL,
  uiContent = NULL,
  external_mod = NULL,
  external_mod_args = list()
)
}
\arguments{
\item{id}{A `character(1)` which is the id of the instance of the module}

\item{title}{A `character(1)`}

\item{width}{A `character(1)` indicating the size of the modal window. Can 
be "s" for small (the default), "m" for medium, or "l" for large.}

\item{uiContent}{The content of the modal dialog.}
}
\value{
A Shiny modal-dialog
}
\description{
Displays of formatted modal-dialog with 'Cancel' and 
'Ok' buttons.
}
