% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{module_exists}
\alias{module_exists}
\title{Checks if a Shiny module is loaded}
\usage{
module_exists(base_name)
}
\arguments{
\item{base_name}{The name of the module (without '_ui' nor '_server' suffixes)}
}
\value{
A boolean
}
\description{
This function checks if the ui() and server() parts of a
Shiny module are available in the global environment.
}
\examples{
path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
source_wf_files(path)
module_exists("PipelineDemo_Process2")

}
