% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{ModuleCoreCode}
\alias{ModuleCoreCode}
\title{Code for declaring xxx}
\usage{
ModuleCoreCode(widgets, steps)
}
\arguments{
\item{widgets}{xxx}

\item{steps}{xxx}
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
steps <- paste0('step', 1:3)
code <- ModuleCoreCode(widgets, steps)
cat(code)
}

}
\author{
Samuel Wieczorek
}
