% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_process.R
\name{nav_process}
\alias{nav_process}
\alias{nav_process_ui}
\alias{nav_process_server}
\title{The server() function of the module `nav_process`}
\usage{
nav_process_ui(id)

nav_process_server(
  id = NULL,
  dataIn = reactive({
     NULL
 }),
  status = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 }),
  remoteReset = reactive({
     0
 }),
  remoteResetUI = reactive({
     0
 }),
  is.skipped = reactive({
     FALSE
 }),
  verbose = FALSE,
  usermod = "user",
  btnEvents = reactive({
     NULL
 })
)

nav_process()
}
\arguments{
\item{id}{A `character(1)` which defines the id of the module.
It is the same as for the server() function.}

\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{status}{xxx}

\item{is.enabled}{A `boolean`. This variable is a remote command to specify
if the corresponding module is enabled/disabled in the calling module of
upper level.
For example, if this module is part of a pipeline and the pipeline calculates
that it is disabled (i.e. skipped), then this variable is set to TRUE. Then,
all the widgets will be disabled. If not, the enabling/disabling of widgets
is deciding by this module.}

\item{remoteReset}{An `integer` which acts as a remote command to reset the 
module. Its value is incremented on a external event and it is used to 
trigger an event in this module}

\item{remoteResetUI}{xxx}

\item{is.skipped}{xxx}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}

\item{usermod}{A character to specifies the running mode of Magellan. 
* user (default) : xxx
* 'dev: xxx}

\item{btnEvents}{xxxx}

\item{history}{A `data.frame` which contains the history of the dataset for
the process of id `id`. If the value is NULL, that means the current process
has not been executed on the dataset, otherwise that means that the process
has already be run on the dataset. This parameter allows to update the
timeline of the process}
}
\value{
A list of four items:
* dataOut A dataset of the same class of the parameter dataIn
* steps.enabled A vector of `boolean` of the same length than config@steps
* status A vector of `integer(1)` of the same length than the config@steps
  vector
* reset xxxx
}
\description{
The module navigation can be launched via a Shiny app.
This is the core module of MagellanNTK
}
\examples{
if (interactive()) {
    nav_process()
}

}
\author{
Samuel Wieczorek
}
