% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_singleProcess.R
\name{nav_single_process}
\alias{nav_single_process}
\alias{nav_single_process_ui}
\alias{nav_single_process_server}
\title{Shiny module `nav_single_process()`}
\usage{
nav_single_process_ui(id)

nav_single_process_server(
  id = NULL,
  dataIn = reactive({
     NULL
 }),
  status = reactive({
     NULL
 }),
  remoteResetUI = reactive({
     0
 }),
  is.enabled = TRUE,
  is.skipped = FALSE,
  verbose = FALSE,
  usermod = "user",
  sendDataIfReset = TRUE
)

nav_single_process()
}
\arguments{
\item{id}{A `character(1)` which defines the id of the module.
It is the same as for the server() function.}

\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{status}{A boolean which indicates whether the current status of the 
process}

\item{remoteResetUI}{xxx}

\item{is.enabled}{A `boolean`. This variable is used as a remote command to specify
if the corresponding module is enabled/disabled in the calling module of
upper level.
For example, if this module is disabled, then this variable is set to TRUE. Then,
all the widgets will be disabled. If not, the enabling/disabling of widgets
is deciding by this module.}

\item{is.skipped}{is.skipped xxx}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}

\item{usermod}{A character to specifies the running mode of MagellanNTK. 
* user (default) : xxx
* dev: xxx}

\item{sendDataIfReset}{A `boolean` to indicate if the reseted value must
be send to the caller in case of reseting the process. This is usefule for example
for the Convert process}
}
\value{
A shiny App
}
\description{
The module navigation can be launched via a Shiny app.
This is the core module of MagellanNTK
}
\examples{
if (interactive()) {
    nav_single_process()
}

}
\author{
Samuel Wieczorek
}
