% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{navbarPageWithInputs}
\alias{navbarPageWithInputs}
\title{xxx}
\usage{
navbarPageWithInputs(..., inputs)
}
\arguments{
\item{...}{xxx}

\item{inputs}{xxx}
}
\description{
Call this function with all the regular navbarPage() parameters,
 plus a text parameter, if you want to add text to the navbar
}
