% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{navbarPageWithText}
\alias{navbarPageWithText}
\title{function for dynamic inputs in DT}
\usage{
navbarPageWithText(..., text)
}
\arguments{
\item{...}{xxx}

\item{text}{xxx}
}
\description{
Call this function with all the regular navbarPage() parameters,
 plus a text parameter, if you want to add text to the navbar
}
