% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{optionsBtnClass}
\alias{optionsBtnClass}
\title{Css class defining the info state}
\format{
An object of class \code{character} of length 1.
}
\usage{
optionsBtnClass
}
\description{
Css class defining the info state
}
\section{Fields}{

\describe{
\item{\code{optionsBtnClass}}{xxxx}
}}

\keyword{datasets}
