% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/utils2.R
\name{pkgs.require}
\alias{pkgs.require}
\title{Loads packages}
\usage{
pkgs.require(ll.deps)

pkgs.require(ll.deps)
}
\arguments{
\item{ll.deps}{A vector of `character()` which contains packages names}
}
\value{
NA
}
\description{
Checks if a package is available to load it

Checks if a package is available to load it
}
\examples{
NULL

\donttest{
pkgs.require(c("stats"))
}
}
\author{
Samuel Wieczorek
}
