% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popover_for_help.R
\name{popover_for_help_ui}
\alias{popover_for_help_ui}
\alias{popover_for_help_server}
\title{popover_for_help_ui and popover_for_help_server}
\usage{
popover_for_help_ui(id)

popover_for_help_server(id, data)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}
}
\value{
xxx
}
\description{
A shiny Module.

xxx
}
\examples{
if (interactive()) {
    library(shiny)
    library(shinyBS)
    ui <- fluidPage(
        popover_for_help_ui("Title")
    )
    server <- function(input, output) {
        popover_for_help_server(
            id = "Title",
            data = list(
                title = "Test",
                content = "Test"
            )
        )
    }
    shinyApp(ui, server)
}

}
