% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{PrepareData2Send}
\alias{PrepareData2Send}
\title{xxxx}
\usage{
PrepareData2Send(rv, pos, verbose = FALSE)
}
\arguments{
\item{rv}{xxxx}

\item{pos}{xxx}

\item{verbose}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
NULL

}
