% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{PrevNextBtnClass}
\alias{PrevNextBtnClass}
\title{Css class defining the previous and next buttons}
\format{
An object of class \code{character} of length 1.
}
\usage{
PrevNextBtnClass
}
\description{
Css class defining the previous and next buttons
}
\section{Fields}{

\describe{
\item{\code{PrevNextBtnClass}}{xxxx}
}}

\keyword{datasets}
