% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layout.R
\name{process_layout}
\alias{process_layout}
\title{Timelines}
\usage{
process_layout(ns, sidebar, content)
}
\arguments{
\item{ns}{xxx}

\item{sidebar}{xxx}

\item{content}{xxx}
}
\value{
NA
}
\description{
xxx
}
