% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_code.R
\name{pull_fifo}
\alias{pull_fifo}
\title{Push item in a FIFO}
\usage{
pull_fifo(fifo)
}
\arguments{
\item{fifo}{A `vector` of strings}
}
\value{
A `list(2)` of two items:
* xxx
* xxx
}
\description{
Push item in a FIFO
}
