% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_code.R
\name{push_fifo}
\alias{push_fifo}
\title{Push item in a FIFO}
\usage{
push_fifo(fifo, value)
}
\arguments{
\item{fifo}{A `vector` of strings}

\item{value}{A string}
}
\value{
A `vector` of strings
}
\description{
Push item in a FIFO
}
