% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readConfigFile}
\alias{readConfigFile}
\title{Read pipelines configuration files}
\usage{
readConfigFile(path, usermod = "user")
}
\arguments{
\item{path}{A `character()` which is the path to the directory which 
contains the files and directories of the pipeline.}

\item{usermod}{A character to specifies the running mode of MagellanNTK. 
* user (default) : xxx
* dev: xxx}
}
\value{
A list containing the following items:
value <- list(
* funcs = tmp,
* verbose = get_data(lines, "verbose") == "enabled",
* UI_view_debugger: a `boolean` to show (TRUE) of hide (FALSE) the debugger interface
* UI_view_open_pipeline: a `boolean` to show (TRUE) of hide (FALSE) the UI to open a pipeline
* UI_view_convert_dataset a `boolean` to show (TRUE) of hide (FALSE) the UI for convert/import a dataset
* UI_view_change_Look_Feel a `boolean` to show (TRUE) of hide (FALSE) the UI to change the L&F
* UI_view_change_core_funcs a `boolean` to show (TRUE) of hide (FALSE) the UI to change the generic functions
* extension A `character()` which specifies the extension file allowed to be processed
in the pipeline,
* class A `character()` which specifies the class of the dataset to be processed
in the pipeline,
* package  A `character()` which specifies the package which owns the pipeline
* demo_package xxx
* URL_manual xxx
* URL_ReleaseNotes xxx
}
\description{
Read the configuration file of a pipeline and extract formatted 
(as a list) information
}
\examples{
path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
readConfigFile(path)

}
