% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCustomizableFuncs}
\alias{readCustomizableFuncs}
\title{Read pipelines configuration files}
\usage{
readCustomizableFuncs(path)
}
\arguments{
\item{path}{xxx}
}
\description{
xxx
}
\examples{
path <- system.file("workflow/PipelineA", package = 'MagellanNTK')
readCustomizableFuncs(file.path(path, 'config.txt'))

}
