% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{redBtnClass}
\alias{redBtnClass}
\title{xxxx}
\format{
An object of class \code{character} of length 1.
}
\usage{
redBtnClass
}
\description{
xxx
}
\details{
#' @title xxxx
#' @description xxx
#' @field btn_style xxxx
#' @export
global = list(VALIDATED = 1,
              SKIPPED = -1,
              UNDONE = 0
)

#' @title xxxx
#' @description xxx
#' @field btn_style xxxx
#' @export
default_pos =list(VALIDATED = 1,
                  SKIPPED = 1,
                  UNDONE = 1
)
}
\section{Fields}{

\describe{
\item{\code{btn_style}}{xxxx}
}}

\keyword{datasets}
