% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\alias{mod_run_workflow_ui}
\alias{mod_run_workflow_server}
\title{A shiny module to run a workflow}
\usage{
mod_run_workflow_ui(id)

mod_run_workflow_server(
  id,
  dataIn = NULL,
  verbose = FALSE,
  tl.layout = NULL,
  path = NULL
)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx
}
\description{
xxxxx
}
\examples{

}
