% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{runDirinputExample}
\alias{runDirinputExample}
\title{Runs a demo app with the code}
\usage{
runDirinputExample()

runDirinputExample()
}
\description{
Runs a demo app with the code

Runs a demo app with the code
}
\details{
Runs a demo app with the code. See the code in the example_application folder for
how this is accomplished.

Runs a demo app with the code. See the code in the example_application folder for
how this is accomplished.
}
