% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_dataset.R
\name{Save_Dataset}
\alias{Save_Dataset}
\alias{Save_Dataset_ui}
\alias{Save_Dataset_server}
\title{Load dataset shiny module}
\usage{
Save_Dataset_ui(id)

Save_Dataset_server(id, data)

Save_Dataset(data)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{data}{The object to be saved}
}
\value{
A shiny app
}
\description{
A shiny Module to load a dataset.
}
\examples{
if (interactive()) {
    shiny::runApp(Save_Dataset(lldata))
}

}
