% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_save.R
\name{Save_Object}
\alias{Save_Object}
\title{Save an R object.}
\usage{
Save_Object(object, file)
}
\arguments{
\item{object}{xxx}

\item{file}{A `characte(1)` naming the file to save the object.}
}
\value{
NA
}
\description{
This function saves an R object to a '.rda' file.
}
\examples{
data(data1)
Save_Object(data1, 'foo.rda')
unlink('foo.rda')

}
