% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ScreenManager.R
\name{ScreenManager}
\alias{ScreenManager}
\title{xxx}
\description{
xxxx
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{xxx}

\item{\code{ns}}{xxx}

\item{\code{verbose}}{xxx}

\item{\code{currentStepName}}{xxx}

\item{\code{child.process}}{xxx}

\item{\code{length}}{xxx}

\item{\code{original.length}}{xxx}

\item{\code{config}}{xxx}

\item{\code{screens}}{xxx}

\item{\code{modal_txt}}{xxx}

\item{\code{timeline}}{xxx}

\item{\code{default_pos}}{xxx}

\item{\code{dataOut}}{xxx}

\item{\code{rv}}{xxx}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScreenManager$new()}}
\item \href{#method-ActionOn_New_DataIn}{\code{ScreenManager$ActionOn_New_DataIn()}}
\item \href{#method-Additional_Server_Funcs}{\code{ScreenManager$Additional_Server_Funcs()}}
\item \href{#method-Set_Skipped}{\code{ScreenManager$Set_Skipped()}}
\item \href{#method-Set_Reseted}{\code{ScreenManager$Set_Reseted()}}
\item \href{#method-ValidateCurrentPos}{\code{ScreenManager$ValidateCurrentPos()}}
\item \href{#method-EncapsulateScreens}{\code{ScreenManager$EncapsulateScreens()}}
\item \href{#method-ActionOn_NewPosition}{\code{ScreenManager$ActionOn_NewPosition()}}
\item \href{#method-Get_Result}{\code{ScreenManager$Get_Result()}}
\item \href{#method-Change_Current_Pos}{\code{ScreenManager$Change_Current_Pos()}}
\item \href{#method-ToggleState_Screens}{\code{ScreenManager$ToggleState_Screens()}}
\item \href{#method-ui}{\code{ScreenManager$ui()}}
\item \href{#method-server}{\code{ScreenManager$server()}}
\item \href{#method-clone}{\code{ScreenManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Change current position.


Default actions on reset pipeline or process.


Set widgets of all screens to their default values.


Gives the name of the status corresponding to the code (integer).


Returns the date and time in timestamp UNIX format.


xxxx


xxx


xxx


xxx


xxx


xxx


Return the UI for a modal dialog with data selection input. If 'failed' is
TRUE, then display a message that the previous value was invalid.


xxx


xxx


xxx


xxx


xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$new(id, verbose)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{xxx}

\item{\code{verbose}}{xxx}

\item{\code{direction}}{xxx}

\item{\code{conf}}{A list containing the configuration of the current object.
See xxx}

\item{\code{name}}{A number}

\item{\code{cond}}{xxx}

\item{\code{range}}{xxx}

\item{\code{cond}}{xxx}

\item{\code{input}}{xxx}

\item{\code{output}}{xxx}

\item{\code{input}}{xxx}

\item{\code{output}}{xxx}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionOn_New_DataIn"></a>}}
\if{latex}{\out{\hypertarget{method-ActionOn_New_DataIn}{}}}
\subsection{Method \code{ActionOn_New_DataIn()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ActionOn_New_DataIn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Additional_Server_Funcs"></a>}}
\if{latex}{\out{\hypertarget{method-Additional_Server_Funcs}{}}}
\subsection{Method \code{Additional_Server_Funcs()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Additional_Server_Funcs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_Skipped"></a>}}
\if{latex}{\out{\hypertarget{method-Set_Skipped}{}}}
\subsection{Method \code{Set_Skipped()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Set_Skipped()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_Reseted"></a>}}
\if{latex}{\out{\hypertarget{method-Set_Reseted}{}}}
\subsection{Method \code{Set_Reseted()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Set_Reseted()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateCurrentPos"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateCurrentPos}{}}}
\subsection{Method \code{ValidateCurrentPos()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ValidateCurrentPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncapsulateScreens"></a>}}
\if{latex}{\out{\hypertarget{method-EncapsulateScreens}{}}}
\subsection{Method \code{EncapsulateScreens()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$EncapsulateScreens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActionOn_NewPosition"></a>}}
\if{latex}{\out{\hypertarget{method-ActionOn_NewPosition}{}}}
\subsection{Method \code{ActionOn_NewPosition()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ActionOn_NewPosition()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Get_Result"></a>}}
\if{latex}{\out{\hypertarget{method-Get_Result}{}}}
\subsection{Method \code{Get_Result()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Get_Result()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Change_Current_Pos"></a>}}
\if{latex}{\out{\hypertarget{method-Change_Current_Pos}{}}}
\subsection{Method \code{Change_Current_Pos()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$Change_Current_Pos(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{xxx}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleState_Screens"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleState_Screens}{}}}
\subsection{Method \code{ToggleState_Screens()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ToggleState_Screens(cond, range)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cond}}{xxx}

\item{\code{range}}{xxx}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
xxx
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$server(dataIn = reactive({
    NULL
}))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataIn}}{xxx}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScreenManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
