% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Send_Result_to_Caller}
\alias{Send_Result_to_Caller}
\title{xxx}
\usage{
Send_Result_to_Caller(data)
}
\arguments{
\item{data}{xxx}
}
\value{
xxx
}
\description{
xxx

xxxx
}
\examples{
Send_Result_to_Caller(3)
}
