% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{SetCurrentPosition}
\alias{SetCurrentPosition}
\title{Get the position of the last validated item}
\usage{
SetCurrentPosition(stepsstatus)
}
\arguments{
\item{stepsstatus}{A vector of integers which reflects the status of the steps 
in the pipeline. Thus, the length of this vector is equal to the number of 
steps}
}
\value{
An integer
}
\description{
Get the position of the last validated item
}
\examples{
status <- c(1,1,1,0,0)
SetCurrentPosition(status)


status <- c(1,1,0,1, 0)
SetCurrentPosition(status)

}
