% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{shinyInput}
\alias{shinyInput}
\title{function for dynamic inputs in DT
#'description xxx}
\usage{
shinyInput(FUN, id, num, ...)
}
\arguments{
\item{FUN}{xxx}

\item{id}{A `character()` as the id of the Shiny module}

\item{num}{The number of widgets}

\item{...}{Additional parameters}
}
\value{
NA
}
\description{
function for dynamic inputs in DT
#'description xxx
}
\examples{
NULL
}
