% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{shinyOutput}
\alias{shinyOutput}
\title{Package version
#'description xxx}
\usage{
shinyOutput(FUN, id, num, ...)
}
\arguments{
\item{FUN}{xxx}

\item{id}{A `character()` as the id of the Shiny module}

\item{num}{The number of widgets}

\item{...}{Additional parameters}
}
\value{
NA
}
\description{
Package version
#'description xxx
}
\examples{
NULL
}
