% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{shinyValue}
\alias{shinyValue}
\title{function to read DT inputs
#'description xxx}
\usage{
shinyValue(id, num)
}
\arguments{
\item{id}{xxx}

\item{num}{xxx}
}
\description{
function to read DT inputs
#'description xxx
}
