% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\name{show,Config-method}
\alias{show,Config-method}
\title{Prints the content of a `Config` class}
\usage{
\S4method{show}{Config}(object)
}
\arguments{
\item{object}{An instance of the class `Config`}
}
\value{
NA
}
\description{
Prints the content of a `Config` class
}
