% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebars}
\alias{sidebars}
\alias{Insert_User_Sidebar}
\title{Sidebar functions}
\usage{
Insert_User_Sidebar()
}
\value{
Shiny UI component
}
\description{
These functions content the UI code for the main sidebar in the
interface of MagellanNTK
}
\examples{
NULL

}
