% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{SimpleWorflowCoreCode}
\alias{SimpleWorflowCoreCode}
\title{Code for declaring xxx}
\usage{
SimpleWorflowCoreCode(name, widgets, steps)
}
\arguments{
\item{name}{xxx}

\item{widgets}{xxx}

\item{steps}{xxx}
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
steps <- paste0('step', 1:3)
code <- SimpleWorflowCoreCode(widgets, steps)
cat(code)
}

}
\author{
Samuel Wieczorek
}
