% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_shinyApp_files}
\alias{source_shinyApp_files}
\title{Source ui.R, server.R ang global.R files}
\usage{
source_shinyApp_files()
}
\value{
NA
}
\description{
Loads the source code of basic Shiny app files
}
\examples{
if (interactive()) {
source_shinyApp_files()
}

}
