% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_wf_files}
\alias{source_wf_files}
\title{Source workflow files}
\usage{
source_wf_files(dirpath, verbose = FALSE)
}
\arguments{
\item{dirpath}{A `character()` which is the path to the directory which 
contains the files and directories of the pipeline.}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}
}
\value{
NA
}
\description{
This function goes into the 'R' directory of the file structure
of a pipeline and load into memory the source code of the R scripts (which contains 
the interfaces of the process)
}
\examples{
path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
source_wf_files(path)

}
