% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_process.R, R/timeline_pipeline.R
\name{timelines}
\alias{timelines}
\alias{timeline_process_ui}
\alias{timeline_process_server}
\alias{timeline_process}
\alias{timeline_pipeline_ui}
\alias{timeline_pipeline_server}
\alias{timeline_pipeline}
\title{Shiny module for the process timeline}
\usage{
timeline_process_ui(id)

timeline_process_server(
  id,
  config,
  status = reactive({
     NULL
 }),
  position = reactive({
     1
 }),
  enabled = reactive({
     NULL
 })
)

timeline_process(config, status, position, enabled)

timeline_pipeline_ui(id)

timeline_pipeline_server(id, config, status, position, enabled)

timeline_pipeline(config, status, position, enabled)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{config}{An instance of the class `Config`}

\item{status}{A boolean which indicates whether the current status of the
pipeline}

\item{position}{An integer which reflects the current position of the cursor
within the steps.}

\item{enabled}{A vector of booleans with the same length as the number of
steps (See the slot steps in the config object). Each element indicates if
the corresponding step is enabled (TRUE) or (DISABLED)}
}
\value{
A shiny app

A shiny App
}
\description{
Define the appearance of the pipeline timeline.

Define the appearance of the pipeline timeline.
}
\examples{
if (interactive()) {
  config <- Config(
    mode = "process",
    fullname = "PipelineDemo_Preprocessing",
    steps = c("Filtering", "Normalization"),
    mandatory = c(FALSE, TRUE)
  )
  status <- reactive({
    c(1, 1, 0, 0)
  })
  pos <- reactive({
    2
  })
  enabled <- reactive({
    c(0, 0, 1, 1)
  })
  shiny::runApp(timeline_process(config, status, pos, enabled))
}

if (interactive()) {
  config <- Config(
    mode = "pipeline",
    fullname = "PipelineDemo",
    steps = c("DataGeneration", "Preprocessing", "Clustering"),
    mandatory = c(TRUE, FALSE, FALSE)
  )
  status <- reactive({
    c(1, 1, -1, 1, 0)
  })
  pos <- reactive({
    4
  })
  enabled <- reactive({
    c(0, 0, 0, 0, 1)
  })
  timeline_pipeline(config, status, pos, enabled)
}

}
