% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Timestamp}
\alias{Timestamp}
\title{xxx}
\usage{
Timestamp()
}
\arguments{
\item{data}{xxx}
}
\value{
xxx
}
\description{
xxx

xxxx

Returns the date and time in timestamp UNIX format.
}
\examples{
Send_Result_to_Caller(3)
}
