% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{ToggleState_NavBtns}
\alias{ToggleState_NavBtns}
\title{Updates the state of navigation buttons}
\usage{
ToggleState_NavBtns(current.pos, nSteps)
}
\arguments{
\item{current.pos}{An `integer` which gives the current position of the cursor
in the corresponding timeline}

\item{nSteps}{The number of steps in the timeline}
}
\value{
NA
}
\description{
Enables/disables the buttons 'Next step' and 'Previous step'
wrt the current position of the cursor in the timeline.
}
\examples{
NULL

}
