% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{ToggleState_ResetBtn}
\alias{ToggleState_ResetBtn}
\title{xxx}
\usage{
ToggleState_ResetBtn(cond)
}
\arguments{
\item{cond}{A boolean: TRUE (enabled), FALSE (disabled)}
}
\value{
NA
}
\description{
See `shinyjs::toggleState()`
}
\examples{
NULL

}
