% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{ToggleState_Screens}
\alias{ToggleState_Screens}
\title{xxx}
\usage{
ToggleState_Screens(cond, range, is.enabled, rv)
}
\arguments{
\item{cond}{A `boolean`}

\item{range}{A `vector` of integers. The min of this vector muste be gerater
of equal to 0 and the max must be less or equal to the size of the vector
rv$steps.enabled}

\item{is.enabled}{A `boolean`}

\item{rv}{A `list` containing at least a slot named 'steps.enabled' which is
a vector of integers}
}
\value{
An updated version of the vector rv$steps.enabled
}
\description{
Updates the status of steps in a given range
}
\examples{
NULL
}
