% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{Update_Data2send_Vector}
\alias{Update_Data2send_Vector}
\title{xxxx}
\usage{
Update_Data2send_Vector(
  current.pos,
  steps.info,
  temp.dataIn,
  dataIn,
  original.length
)
}
\arguments{
\item{current.pos}{xxx}

\item{steps.info}{xxx}

\item{temp.dataIn}{xxx}

\item{dataIn}{xxx}

\item{original.length}{xxx}
}
\description{
xxx

xxx
}
