% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{UpdateStepsStatus}
\alias{UpdateStepsStatus}
\title{xxx}
\usage{
UpdateStepsStatus(dataIn, config)
}
\arguments{
\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{config}{An instance of the `Config` class}
}
\value{
A vector of boolean
}
\description{
xxx
}
\examples{
NULL
}
