% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_dataset.R
\name{view_dataset}
\alias{view_dataset}
\alias{view_dataset_ui}
\alias{view_dataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
view_dataset_ui(id)

view_dataset_server(id, dataIn = NULL, ...)

view_dataset(dataIn)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{dataIn}{An instance of the class `MultiAssayExperiment`}

\item{...}{Additional parameters}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
  data(lldata123)
  view_dataset(lldata123)
}

}
