% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{workflow}
\alias{workflow}
\alias{pipe_workflow_ui}
\alias{pipe_workflow_server}
\alias{pipe_workflowApp}
\alias{proc_workflow_ui}
\alias{proc_workflow_server}
\alias{proc_workflowApp}
\title{Shiny example module `Process`}
\usage{
pipe_workflow_ui(id)

pipe_workflow_server(
  id,
  path = NULL,
  dataIn = reactive({
     NULL
 }),
  usermod = "user",
  verbose = FALSE
)

pipe_workflowApp(
  id,
  path = NULL,
  dataIn = NULL,
  usermod = "user",
  verbose = FALSE
)

proc_workflow_ui(id)

proc_workflow_server(
  id,
  path = NULL,
  dataIn = reactive({
     NULL
 }),
  usermod = "user",
  verbose = FALSE
)

proc_workflowApp(
  id,
  path = NULL,
  dataIn = NULL,
  usermod = "user",
  verbose = FALSE
)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{path}{A `character()` which is the path to the directory which 
contains the files and directories of the pipeline.}

\item{dataIn}{An instance of the class `MultiAssayExperiment`}

\item{usermod}{A `character()` to specifies the running mode of MagellanNTK: 
'user' (default) or 'dev'. For more details, please refer to the document 
'Inside MagellanNTK'}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}
}
\value{
A shiny app

A shiny App
}
\description{
This module contains the configuration information for the corresponding
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developers who want to create their own
pipelines nor processes to be managed with `MagellanNTK`.
}
\examples{
if (interactive()) {
    library(MagellanNTK)
    library(shiny)
    path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
    files <- list.files(file.path(path, "R"), full.names = TRUE)
    for (f in files) {
        source(f, local = FALSE, chdir = TRUE)
    }


    # Nothing happens when dataIn is NULL
    pipe_workflowApp("PipelineDemo", path, dataIn = NULL)

    pipe_workflowApp("PipelineDemo", path, dataIn = lldata)

    pipe_workflowApp("PipelineDemo", dataIn = data.frame())

    pipe_workflowApp("PipelineDemo", path, dataIn = lldata)

    pipe_workflowApp("PipelineB", path, tl.layout = c("v", "h"))

    pipe_workflowApp("PipelineDemo", path,
        dataIn = lldata,
        tl.layout = c("v", "h")
    )


    # Nothing happens when dataIn is NULL
    proc_workflowApp("PipelineDemo_Process1", path, dataIn = NULL)

    proc_workflowApp("PipelineDemo_Process1", path, dataIn = lldata)

    proc_workflowApp("PipelineDemo_Process1", path, dataIn = data.frame(), tl.layout = "v")

    proc_workflowApp("PipelineDemo", path, dataIn = lldata, tl.layout = c("v", "h"))

    proc_workflowApp("PipelineDemo", path, dataIn = lldata, tl.layout = c("v", "h"))
}

}
\author{
Samuel Wieczorek
}
