% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write.excel}
\alias{write.excel}
\title{This function exports a data.frame to a Excel file.}
\usage{
write.excel(df, colors, tags, filename = NULL)
}
\arguments{
\item{df}{A list of data.frame() items}

\item{colors}{xxx}

\item{tags}{xxx}

\item{filename}{A character string for the name of the Excel file.}
}
\value{
A Excel file (.xlsx)
}
\description{
This is the default function fo export the current dataset as an Excel file.
}
\examples{
data(lldata)
write.excel(lldata, filename = "foo.xlsx")

}
\author{
Samuel Wieczorek
}
