% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{ActionOn_Child_Changed}
\alias{ActionOn_Child_Changed}
\title{xxx}
\usage{
ActionOn_Child_Changed(
  temp.dataIn,
  dataIn,
  steps.status,
  steps,
  steps.enabled,
  steps.skipped,
  processHasChanged,
  newValue
)
}
\arguments{
\item{dataIn}{xxx}

\item{steps.status}{xxx}

\item{steps}{xxx}

\item{steps.enabled}{xxx}

\item{steps.skipped}{xxx}

\item{processHasChanged}{xxxx}

\item{newValue}{xxx}
}
\description{
# A process has changed
}
\author{
Samuel Wieczorek
}
