% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{Add2History}
\alias{Add2History}
\title{Get the last validated step before current position.}
\usage{
Add2History(history, process, step.name, param.name, value)
}
\arguments{
\item{history}{A `data.frame()`}

\item{process}{A `character()`}

\item{step.name}{A `character()`}

\item{param.name}{A `character()`}

\item{value}{The value corresponding to the param.name}
}
\value{
A `data.frame()`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
history <- InitializeHistory()
Add2History(history, 'Example', 'First step', "my param", 'THE value')


}
