% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{Add2History}
\alias{Add2History}
\title{Get the last validated step before current position.}
\usage{
Add2History(history, process, step.name, param.name, value)
}
\arguments{
\item{widgets.names}{xxx}
}
\value{
A `integer(1)`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL
}
