% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pipeline.R, R/mod_nav_pipeline.R
\name{Add_Item_to_Dataset}
\alias{Add_Item_to_Dataset}
\title{xxx}
\usage{
Add_Item_to_Dataset(dataset, name)

Add_Item_to_Dataset(dataset, name)
}
\arguments{
\item{dataset}{xxx}

\item{name}{xxx}
}
\value{
The dataset minus some items

The dataset minus some items
}
\description{
Adds one or more items to the dataset. This function is specific of the
type of dataset.

Adds one or more items to the dataset. This function is specific of the
type of dataset.
}
