% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_build_report.R
\name{build_report}
\alias{build_report}
\alias{build_report_ui}
\alias{build_report_server}
\title{dl}
\usage{
build_report_ui(id)

build_report_server(
  id,
  dataIn = reactive({
     NULL
 }),
  filename = "myDataset"
)
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{filename}{xxx}

\item{widget.type}{Available values are `Button` and `Link` (default).}

\item{name}{internal}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
\dontrun{
data(sub_R25)
shiny::runApp(build_report(sub_R25))

shiny::runApp(build_report(sub_R25, filename = 'myDataset'))
}

}
