% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{call.func}
\alias{call.func}
\title{Call function}
\usage{
call.func(fname, args)
}
\arguments{
\item{fname}{xxx}

\item{args}{xxx}
}
\description{
xxx
}
\examples{
call.func("stats::rnorm", list(10))

}
\seealso{
[do.call()]
}
