% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckConfig}
\alias{CheckConfig}
\title{xxx}
\usage{
CheckConfig(config)
}
\arguments{
\item{config}{A list containing the rv$configuration of the current object.
See xxx}
}
\value{
A list of two items:
* `passed`: A boolean that indicates if the config is correct or not.
* `msg`: A `character(1)` as message.
}
\description{
xxx

Check if the rv$config is correct
}
\author{
Samuel Wieczorek
}
