% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chooseDir.R
\name{choose_dir}
\alias{choose_dir}
\alias{chooseDir_ui}
\alias{chooseDir_server}
\alias{chooseDir}
\title{Module choose directory}
\usage{
chooseDir_ui(id)

chooseDir_server(
  id,
  path = reactive({
     "~"
 }),
  is.enabled = reactive({
     TRUE
 })
)

chooseDir()
}
\arguments{
\item{id}{xxx}

\item{path}{xxx}

\item{is.enabled}{xxx}
}
\description{
A shiny Module which xxx
}
\examples{
NULL

}
\author{
Samuel Wieczorek
}
