% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{CleanSourceCode}
\alias{CleanSourceCode}
\title{Clean source code before syntax analysis}
\usage{
CleanSourceCode(file = NULL)
}
\arguments{
\item{file}{A vector in which each element is a line read from source
code file}
}
\value{
A string which contains the source code condensed
}
\description{
xxx
}
\examples{
file <- system.file("extdata/module_examples", "mod_Process1.R", package = "Magellan")
CleanSourceCode(file)

}
\author{
Samuel Wieczorek
}
