% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\name{Config-class}
\alias{Config-class}
\alias{Config}
\alias{initialize,Config-method}
\title{Config class definition}
\usage{
Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)

\S4method{initialize}{Config}(.Object, fullname, mode, steps, mandatory, steps.source.file)

Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)
}
\arguments{
\item{fullname}{The complete name of the pipeline or the process (in this case,
it is the concatenation of the name of the pipeline and the name of the process 
itself, separated by '_')}

\item{mode}{A `character()` which indicates whether the current module is used
as a 'process' nor a 'pipeline.' Default value is NULL}

\item{steps}{A vector containing the names of steps in a process or a pipeline.}

\item{mandatory}{A vector of boolean where each element indicates whether the
corresponding steps is mandatory or not. It has the same length of the vector
steps.}

\item{steps.source.file}{A `vector` of which each item is the source code
of the corresponding step.}

\item{.Object}{xxx}
}
\value{
NA
}
\description{
This class is used to store the configuration of any process
used with MagellanNTK It contains a validity function to ensure
that the format is correct.

Validity:
* The first step must be called 'Description', it is a mandatory step. Thus,
the first item of the mandatory vector is TRUE.
To be continued...

## Initialization
 ### Generic process

 A generic process
 * Generic pipeline : xxxx
 * Description pipeline: This case is for a process called 'Description' which is
 the first process module of a pipeline

Wrapper function to the constructor of the class
}
\section{Slots}{

\describe{
\item{\code{fullname}}{The name of the process which is the concatenation of the
name of the pipeline and the name of the process itself, separated by '_'.}

\item{\code{name}}{The name of the process nor pipeline.}

\item{\code{parent}}{The name of the pipeline/process which owns this instance}

\item{\code{mode}}{A `character()` which indicates if the configuration is about
a whole process nor a process of a pipeline.}

\item{\code{steps}}{A `vector` of `character()` which contains the primary steps
of the pipeline. Two steps will be automatically inserted in this vector:
'Description 'in the first position and 'Save' at the end.}

\item{\code{mandatory}}{A `vector` of `boolean()` in which each item is the 
necessary code for the GUI of a step. The size of this vector is the number
 of steps}

\item{\code{ll.UI}}{A `vector` of Shiny source code. Each item is the necessary code 
for the GUI of a step. The size of this vector is the number of steps}

\item{\code{steps.source.file}}{A `vector` of which each item is the source code
of the corresponding step.}

\item{\code{steps}}{A `vector` of `character()` which contains the primary steps
of the pipeline. Two steps will be automatically inserted in this vector:
'Description 'in the first position and 'Save' at the end.}

\item{\code{mandatory}}{A `vector` of `boolean()` in which each item is the 
necessary code for the GUI of a step. The size of this vector is the number
 of steps}
}}

\examples{

# Example of a generic process
generic.proc <- Config(
    fullname = "PipelineDemo_Process1",
    mode = "process",
    steps = c("Step 1", "Step 2"),
    mandatory = c(TRUE, FALSE)
)


# Example of a generic pipeline
generic.pipe <- Config(
    fullname = "Pipe1_PipelineDemo",
    mode = "pipeline",
    steps = c("Process 1", "Process 2"),
    mandatory = c(TRUE, FALSE)
)

# Example of a root pipeline (process has no parent)
root.pipe <- Config(
    mode = "pipeline",
    fullname = "PipelineDemo",
    steps = c("Process1", "Process2 bis", "Process3"),
    mandatory = c(FALSE, FALSE, TRUE)
)


# Example of a description module (process has no steps)
description.process <- Config(
    fullname = "PipelineDemo_Description",
    mode = "process",
    steps = "",
    mandatory = ""
)


generic.proc
generic.pipe
root.pipe
description.process

}
