% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\name{Config-class}
\alias{Config-class}
\alias{Config}
\alias{initialize,Config-method}
\title{Initialization method for the class `Config`}
\usage{
Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)

\S4method{initialize}{Config}(.Object, fullname, mode, steps, mandatory, steps.source.file)

Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)
}
\arguments{
\item{fullname}{xxx}

\item{mode}{xxx}

\item{steps}{xxx}

\item{mandatory}{xxx}

\item{steps.source.file}{xxx}

\item{.Object}{xxx}
}
\description{
Wrapper function to the constructor of the class
}
