% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\docType{class}
\name{Config-class}
\alias{Config-class}
\alias{Config}
\title{Config class definition}
\description{
This class is used to store the configuration of any process
used with Magellan. It contains a validity function to ensure
that the format is correct.
}
