% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_class.R
\docType{class}
\name{Config-class}
\alias{Config-class}
\alias{Config}
\title{Config class definition}
\description{
This class is used to store the configuration of any process
used with MagellanNTK It contains a validity function to ensure
that the format is correct.

Validity:
* The first step must be called 'Description', it is a mandatory step. Thus, 
the first item of the mandatory vector is TRUE.
To be continued...
}
\examples{
config <- Config(
mode = 'pipeline',
name = 'PipelineA',
# List of all steps of the process
# Here, each step is a workflow
steps = c('Description', 'Process1', 'Process2', 'Process3'),
# A vector of boolean indicating if the steps are mandatory or not.
mandatory = c(TRUE, FALSE, FALSE, TRUE),

path_to_md_file = system.file('extdata/module_examples/md/PipelineA.md', package='MagellanNTK')
)


}
