% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMdDescriptionFiles.R
\name{Create_md_file}
\alias{Create_md_file}
\title{xxx}
\usage{
Create_md_file(mode, fullname, path, rawText)
}
\arguments{
\item{mode}{xxx}

\item{fullname}{xxx}

\item{path}{xxx}

\item{rawText}{xxx}
}
\description{
xxx
}
