% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModuleTemplate.R
\name{createModuleTemplate}
\alias{createModuleTemplate}
\title{Create process template code}
\usage{
createModuleTemplate(config = NULL, path = ".")
}
\arguments{
\item{config}{An instance of the class `Config`.}

\item{path}{xxx}
}
\value{
NA
}
\description{
This function creates the source code of an empty module 
(i.e. a module without any widgets). This module contains the minimal
skeleton to work. The developer can then insert its own code for widgets
and data processing functions.

The 'Description' step is generic and creates a *.md file to be filled by
th developer.
}
\examples{
if(interactive()){
  
  conf.process <- Config(
  name = 'ProcessX',
  mode = "process",
  steps = c('Description', "Step 1", "Step 2", "Save"),
  mandatory = c(TRUE, TRUE, FALSE, TRUE),
    path_to_md_file = '.'
  )


file <- createModuleTemplate(conf.process)
source(file)
run_workflow("ProcessX", verbose = TRUE)
unlink(file)
  
  
conf.pipeline <- Config(
  mode = "pipeline",
  name = 'PipelineA',
  steps = c('Description', "Process 1", "Process 2", "Process 3"),
  mandatory = c(TRUE, TRUE, FALSE, TRUE),
  path_to_md_file = system.file("extdata/module_examples/md/", package = "MagellanNTK")
)
createModuleTemplate(conf.pipeline)
run_workflow("PipelineA", verbose = TRUE)

}
}
\author{
Samuel Wieczorek
}
