% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModuleTemplate.R
\name{createTemplate}
\alias{createTemplate}
\alias{createModuleTemplate}
\alias{write_Code_for_Description_file}
\alias{write_process_ui_function}
\alias{write_process_header_server_func}
\alias{vec2code}
\alias{write_get_process_config_code}
\alias{write_process_code_for_default_value_widgets}
\alias{write_process_module_server_header}
\alias{write_process_renderUI_for_steps}
\alias{write_Insert_Description_Step_code_for_Process}
\alias{write_Insert_Description_Step_code_for_Pipeline}
\alias{write_process_output_func}
\alias{write_pipeline_ui_function}
\alias{write_pipeline_header_server_func}
\alias{write_pipeline_config_code}
\alias{write_pipeline_module_server}
\title{Create process template code}
\usage{
createModuleTemplate(config = NULL, path = ".")

write_Code_for_Description_file(con, name)

write_process_ui_function(con, name)

write_process_header_server_func(con, name)

vec2code(ls_list, is.char = FALSE)

write_get_process_config_code(con, config)

write_process_code_for_default_value_widgets(con)

write_process_module_server_header(con)

write_process_renderUI_for_steps(con, steps)

write_Insert_Description_Step_code_for_Process(con)

write_Insert_Description_Step_code_for_Pipeline(con)

write_process_output_func(con)

write_pipeline_ui_function(con, name)

write_pipeline_header_server_func(con, name)

write_pipeline_config_code(con, config)

write_pipeline_module_server(con)
}
\arguments{
\item{config}{An instance of the class `Config`.}

\item{path}{xxx}

\item{name}{xxx}

\item{ls_list}{A vector}

\item{is.char}{A `bolean(1)` to indicate whether the items are
strings or not. In this case, they will be quoted in the result}

\item{steps}{xxx}
}
\value{
NA

A string
}
\description{
This function creates the source code of an empty module 
(i.e. a module without any widgets). This module contains the minimal
skeleton to work. The developer can then insert its own code for widgets
and data processing functions.

The 'Description' step is generic and creates a *.md file to be filled by
th developer.

This function inserts the necessary code for the 'Description' step

This function inserts the necessary code for the 
Description' step
}
\examples{
if(interactive()){
  
  conf.process <- Config(
    full.name = 'PipeA_ProcessX',
    mode = "process",
    steps = c("Step 1", "Step 2"),
    mandatory = c(TRUE, FALSE)
    )


file <- createModuleTemplate(conf.process)
source(file)
run_workflow("ProcessX", verbose = TRUE)
unlink(file)
  
  
conf.pipeline <- Config(
  mode = "pipeline",
  full.name = 'PipelineA',
  steps = c("Process 1", "Process 2", "Process 3"),
  mandatory = c(TRUE, FALSE, TRUE)
  )
createModuleTemplate(conf.pipeline)
run_workflow("PipelineA", verbose = TRUE)

}
}
\author{
Samuel Wieczorek
}
