% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{Keep_Datasets_from_Object}
\alias{Keep_Datasets_from_Object,list-method}
\alias{Keep_Datasets_from_Object,QFeatures-method}
\alias{Add_Datasets_to_Object}
\alias{Add_Datasets_to_Object,list-method}
\alias{Add_Datasets_to_Object,QFeatures-method}
\title{Datasets processing}
\usage{
\S4method{Keep_Datasets_from_Object}{list}(object, range)

\S4method{Add_Datasets_to_Object}{list}(object, dataset, name)
}
\arguments{
\item{object}{An object of class `list`.}

\item{range}{A xxxx}

\item{dataset}{`character(1)` providing the base with respect to which
logarithms are computed. Defaults is 2.}

\item{name}{A `character(1)` naming the new assay name.}
}
\value{
An processed object of the same class as `object`.
}
\description{
This manual page describes manipulation methods using [list] objects. In the following
functions, if `object` is of class `list`, and optional assay
index or name `i` can be specified to define the assay (by name of
index) on which to operate.

The following functions are currently available:

- `Keep_Datasets_from_Object(object, range)` keep datasets in object which
are in range

- `Add_Datasets_to_Object(object, dataset, name)` add the 'dataset' to the object (of type list)
}
\details{
The object must be of type list. Thetwo functions are implemented here for a simple list.
For other dataset classes, their implementation must be part of the package which uses Magellan.
}
\examples{

}
