% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataset-processing}
\alias{dataset-processing}
\alias{Keep_Datasets_from_Object}
\alias{Keep_Datasets_from_Object,list-method}
\alias{Add_Datasets_to_Object}
\alias{Add_Datasets_to_Object,list-method}
\alias{Keep_Datasets_from_Object,NULL-method}
\alias{Add_Datasets_to_Object,NULL-method}
\title{Datasets processing}
\usage{
\S4method{Keep_Datasets_from_Object}{`NULL`}(object, range)

\S4method{Keep_Datasets_from_Object}{list}(object, range)

\S4method{Add_Datasets_to_Object}{`NULL`}(object, dataset, name)

\S4method{Add_Datasets_to_Object}{list}(object, dataset, name)
}
\arguments{
\item{object}{An object of class `list`.}

\item{range}{A xxxx}

\item{dataset}{`character(1)` providing the base with respect to which
logarithms are computed. Default is log2.}

\item{name}{A `character(1)` naming the new array name.}
}
\value{
An processed object of the same class as `object`.
}
\description{
This manual page describes manipulation methods using [list] objects. In 
index or name `i` can be specified to define the array (by name of
index) on which to operate.

The following functions are currently available:

- `Keep_Datasets_from_Object(object, range)` keep datasets in object which
are in range

- `Add_Datasets_to_Object(object, dataset, name)` add the 'dataset' to the 
object (of type list)

- `Save(object, file)` stores the object to a .RData file
}
\details{
The object must be of type list. Thetwo functions are implemented here for 
of the package which uses Magellan.
}
\examples{
NA

}
