% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{data1}
\alias{data_na}
\title{data1 dataset}
\format{
An object of class `list()`

An object of class `list()`
}
\usage{
data(data1)

data(data_na)
}
\description{
xxx

xxx
}
\keyword{data}
\keyword{datasets}
