% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_debug_infos.R
\name{Debug_Infos}
\alias{Debug_Infos}
\alias{Debug_Infos_ui}
\alias{Debug_Infos_server}
\title{Module debugging}
\usage{
Debug_Infos_ui(id)

Debug_Infos_server(
  id,
  title = NULL,
  config = reactive({
     NULL
 }),
  rv.dataIn = reactive({
     NULL
 }),
  dataIn = reactive({
     NULL
 }),
  dataOut = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  steps.status = reactive({
     NULL
 }),
  steps.skipped = reactive({
     NULL
 }),
  current.pos = reactive({
     NULL
 }),
  is.enabled = reactive({
     NULL
 })
)

Debug_Infos(obj)
}
\arguments{
\item{id}{xxx}

\item{title}{The title of the Panel which contains the debugging tables}

\item{config}{An instance of the class `Config`}

\item{rv.dataIn}{xxx}

\item{dataIn}{A `list()` of data.frames}

\item{dataOut}{A `list()` of data.frames}

\item{steps.enabled}{A `logical()` xxxx}

\item{steps.status}{A `logical()` xxxx}

\item{steps.skipped}{A `logical()` xxxx}

\item{current.pos}{A `integer(1)` which is the indice of the active step.}

\item{is.enabled}{A `logical(1)` xxxx}

\item{obj}{xxx}
}
\value{
NA

A shiny app
}
\description{
A shiny Module which shows the values of the datasets variables
along the different processes of a process nor pipeline.
}
\examples{
\dontrun{
shiny::runApp(Debug_Infos(sub_R25))
}

}
\author{
Samuel Wieczorek
}
