% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_default_plugins.R
\name{default_plugin}
\alias{default_plugin}
\alias{mod_openfile_ui}
\alias{mod_openfile_server}
\alias{mod_plots_ui}
\alias{mod_plots_server}
\alias{mod_export_ui}
\alias{mod_export_server}
\alias{mod_convert_ui}
\alias{mod_convert_server}
\title{xxxx}
\usage{
mod_openfile_ui(id)

mod_openfile_server(
  id,
  path = reactive({
     NULL
 })
)

mod_plots_ui(id)

mod_plots_server(id, object)

mod_export_ui(id)

mod_export_server(id, object)

mod_convert_ui(id)

mod_convert_server(id, object)
}
\arguments{
\item{id}{xxxx}

\item{path}{xxx}
}
\description{
xx
}
