% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_default_plugins.R
\name{default_plugin}
\alias{default_plugin}
\alias{default_openfile_ui}
\alias{default_openfile_server}
\alias{default_EDA_ui}
\alias{default_EDA_server}
\alias{default_export_ui}
\alias{default_export_server}
\title{xxxx}
\usage{
default_openfile_ui(id)

default_openfile_server(
  id,
  path = reactive({
     NULL
 })
)

default_EDA_ui(id)

default_EDA_server(id, object)

default_export_ui(id)

default_export_server(id, object)
}
\arguments{
\item{id}{xxxx}

\item{path}{xxx}
}
\description{
xx
}
