% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_variables.R
\name{default_vars}
\alias{default_vars}
\alias{default.funcs}
\alias{default.base.URL}
\alias{default.workflow}
\title{Default vars}
\usage{
default.funcs()

default.base.URL()

default.workflow()
}
\description{
Default vars
}
