% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_variables.R
\docType{data}
\name{default_vars}
\alias{default_vars}
\alias{default.funcs}
\alias{default.base.URL}
\alias{default.workflow}
\alias{default.layout}
\title{Default vars}
\format{
An object of class \code{list} of length 7.
}
\usage{
default.funcs()

default.base.URL()

default.workflow()

default.layout
}
\description{
Default vars
}
\keyword{datasets}
