% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_variables.R
\docType{data}
\name{default_vars}
\alias{default_vars}
\alias{default.funcs}
\alias{default.base.URL}
\alias{default.workflow}
\alias{default.layout}
\alias{default.theme}
\title{Default vars}
\format{
An object of class \code{list} of length 37.
}
\usage{
default.funcs()

default.base.URL()

default.workflow()

default.layout

default.theme(mode)
}
\arguments{
\item{mode}{xxx}
}
\value{
NA
}
\description{
left_panel = left_sidebar + width_sidebar
top_sidebar = top_panel
width_sidebar = width_process_btns

left_panel = left_sidebar + width_sidebar
top_sidebar = top_panel
width_sidebar = width_process_btns
}
\examples{
NULL
}
\keyword{datasets}
