% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dl.R, R/mod_download.R
\name{dl}
\alias{dl}
\alias{dl_ui}
\alias{dl_server}
\alias{download_ui}
\alias{download_server}
\title{dl}
\usage{
dl_ui(id)

dl_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = "csv",
  widget.type = "Link",
  name = "foo",
  excel.style = NULL
)

download_ui(id)

download_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = reactive({
     "csv"
 }),
  widget.type = reactive({
     "Link"
 }),
  name = reactive({
     "foo"
 }),
  excel.style = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are `csv` (default), `RData` and `Excel`.}

\item{widget.type}{Available values are `Button` and `Link` (default).}

\item{name}{internal}

\item{excel.style}{xxx}
}
\value{
NA

NA
}
\description{
A shiny Module.

A shiny Module.
}
\examples{
NULL


}
