% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_dataset_ui}
\alias{download_dataset_server}
\title{dl}
\usage{
download_dataset_ui(id)

download_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  filename = "myDataset",
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

download_dataset(dataIn = NULL, filename = "myDataset")
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{dataIn}{An instance of the class `MultiAssayExperiment`}

\item{filename}{internal}

\item{remoteReset}{A `logical(1)` which acts as a remote command to reset
the module to its default values. Default is FALSE.}

\item{is.enabled}{A `boolean`. This variable is used as a remote command to specify
if the corresponding module is enabled (TRUE) or disabled (FALSE). For more
details, please refer to the document 'Inside MagellanNTK'?}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
data(lldata, package = "MagellanNTK")
shiny::runApp(download_dataset(lldata))
}


}
