% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_dataset_ui}
\alias{download_dataset_server}
\title{dl}
\usage{
download_dataset_ui(id)

download_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  widget.type = "Link",
  filename = "myDataset"
)

download_dataset(data, filename = "myDataset")
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{dataIn}{internal}

\item{widget.type}{Available values are `Button` and `Link` (default).}

\item{filename}{xxx}

\item{data}{xxx}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(download_dataset(lldata))

    shiny::runApp(download_dataset(lldata, filename = "myDataset"))
}

}
