% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\alias{download_dataset_ui}
\alias{download_dataset_server}
\title{dl}
\usage{
download_dataset_ui(id)

download_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  extension = c("csv", "xlsx", "RData"),
  widget.type = "Link",
  name = "foo",
  excel.style = NULL
)

download_dataset(data)
}
\arguments{
\item{id}{internal}

\item{dataIn}{internal}

\item{extension}{Available values are `csv` (default), `RData` and `Excel`.}

\item{widget.type}{Available values are `Button` and `Link` (default).}

\item{name}{internal}

\item{excel.style}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(sub_R25)
shiny::runApp(download_dataset(sub_R25))
\dontshow{\}) # examplesIf}
}
