% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Shiny example module `Process`}
\usage{
run_workflow(id, verbose = FALSE, tl.layout = NULL)
}
\arguments{
\item{id}{xxx}

\item{verbose}{A `boolean` that indicates whether to show some infos in the 
console and add the shiny module for debugging}

\item{tl.layout}{Additional parameters for mod_nav}
}
\value{
NA
}
\description{
This module contains the configuration information for the corresponding 
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developers who want to create their own 
pipelines nor processes to be managed with `MagellanNTK`.
}
\examples{
if (interactive()) {
    dirpath <- system.file("extdata/module_examples", package = "MagellanNTK")
    for (l in list.files(path = dirpath, pattern = ".R", recursive = TRUE)) {
        source(file.path(dirpath, l), local = FALSE)$value
    }

    run_workflow("Process1", verbose = TRUE)

    run_workflow("PipelineA", tl.layout = c("h", "h"))
}
}
\author{
Samuel Wieczorek
}
