% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_createTemplate.R, R/run_workflow.R
\name{Tools_Templates}
\alias{Tools_Templates}
\alias{demo_workflow}
\title{xxx}
\usage{
Tools_Templates()

demo_workflow(name = NULL, dataIn = NULL)
}
\arguments{
\item{name}{xxx}
}
\description{
xxx

xxx
}
