% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Shiny example module `Process`}
\usage{
run_workflow(id, dataIn = NULL, tl.layout = NULL, path = NULL, mode = "user")
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{tl.layout}{Additional parameters for nav}

\item{path}{The path to the directory where are the source code files 
for processes and pipelines}

\item{mode}{xxx}
}
\value{
NA
}
\description{
This module contains the configuration information for the corresponding 
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developers who want to create their own 
pipelines nor processes to be managed with `MagellanNTK`.
}
\examples{
if (interactive()) {
  data(data_na)
  
  path <- system.file("extdata/module_examples", package = "MagellanNTK")
  
  # Nothing happens whien dataIn is NULL
  run_workflow("PipelineA_Process1", dataIn = NULL, path = path)

  run_workflow("PipelineA", dataIn = data_na, path = path, tl.layout = c("v", "h"))
    
  run_workflow("PipelineB", path = path, tl.layout = c("v", "h"))
}
}
\author{
Samuel Wieczorek
}
