% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_funcs.R
\name{FindClosingParenthesis}
\alias{FindClosingParenthesis}
\title{xxx}
\usage{
FindClosingParenthesis(text, openPos)
}
\arguments{
\item{text}{xxx}

\item{openPos}{xxx}
}
\description{
xxx
}
\examples{
text <- "myfunc <- function(a, b){ return (a+b)}"
posParam <- 19
FindClosingParenthesis(text, posParam)

}
