% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{format_DT}
\alias{format_DT}
\alias{format_DT_ui}
\alias{format_DT_server}
\title{format_DT_ui and format_DT_server}
\usage{
format_DT_ui(id)

format_DT_server(id, obj)

format_DT(obj)
}
\arguments{
\item{id}{shiny id}

\item{obj}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(lldata)
shiny::runApp(format_DT(assay(lldata[[1]])))

\dontshow{\}) # examplesIf}
}
