% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{generic_mod_open_dataset}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{Change the default functions in `MagellanNTK`}
\usage{
open_dataset_ui(id)

open_dataset_server(
  id,
  class = NULL,
  extension = NULL,
  demo_package = NULL,
  remoteReset = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 })
)

open_dataset(class = NULL, extension = NULL, demo_package = NULL)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{class}{xxx}

\item{extension}{xxx}

\item{demo_package}{xxx}

\item{remoteReset}{An `integer` which acts as a remote command to reset the 
module. Its value is incremented on a external event and it is used to 
trigger an event in this module}

\item{is.enabled}{xxx}
}
\value{
NA
}
\description{
This module allows to change
}
\examples{
if (interactive()) {
shiny::runApp(open_dataset(extension = "df"))
}

}
