% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{generic_mod_open_dataset}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{Change the default functions in `MagellanNTK`}
\usage{
open_dataset_ui(id)

open_dataset_server(id, class = NULL, extension = NULL, demo_package = NULL)

open_dataset(class = NULL, extension = NULL, demo_package = NULL)
}
\arguments{
\item{id}{xxx}

\item{class}{xxx}

\item{extension}{xxx}

\item{demo_package}{xxx}
}
\value{
NA
}
\description{
This module allows to change
}
\examples{
if (interactive()) {
    shiny::runApp(open_dataset(class = "QFeatures", extension = ".qf"))
    shiny::runApp(open_dataset(class = "QFeatures", extension = ".qf", demo_package = "DaparToolshedData"))
}

}
