% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_launch_magellan.R, R/mod_open_dataset.R
\name{mod_launch_magellan}
\alias{mod_launch_magellan}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{Change the default functions in `MagellanNTK`}
\usage{
mod_launch_magellan()

open_dataset_ui(id)

open_dataset_server(id, class = NULL, extension = NULL, demo_package = NULL)

open_dataset(class = NULL, extension = NULL)
}
\arguments{
\item{id}{xxx}
}
\description{
This module allows to change
}
\examples{
\dontrun{
shiny::runApp(open_dataset(class = 'QFeatures', extension = '.qf'))
}


}
