% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{generic_mod_open_dataset}
\alias{generic_mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{mod_open_dataset_ui and mod_open_dataset_server}
\usage{
open_dataset_ui(id)

open_dataset_server(id)

open_dataset()
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(open_dataset())


\dontshow{\}) # examplesIf}
}
\keyword{internal}
