% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_AdditionalModule_Core_Code}
\alias{Get_AdditionalModule_Core_Code}
\title{Code for declaring xxx}
\usage{
Get_AdditionalModule_Core_Code(w.names, rv.custom.names = NULL)
}
\arguments{
\item{w.names}{xxx}

\item{rv.custom.names}{xxx}
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
steps <- paste0('step', 1:3)
code <- SimpleWorflowCoreCode(widgets, steps)
cat(code)
}

}
\author{
Samuel Wieczorek
}
