% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_external_modules.R
\name{Get_Code_Declare_widgets}
\alias{Get_Code_Declare_widgets}
\title{Code for declaring widgets.default.values reactive variable}
\usage{
Get_Code_Declare_widgets(widgets.names = NULL)
}
\arguments{
\item{widgets.names}{A `list` containing the names of the widgets in all
steps of the module.}
}
\value{
NA
}
\description{
This function create the source code needed inside a module
to declare the reactive variable called 'widgets.default.values'.
# Declaration of the variables that will contain the values of the widgets
To avoid confusion, the first string is the name of the step while the
second is the name of the widget
}
\examples{
widgets <- paste0("widget", 1:3)
Get_Code_Declare_widgets(widgets)

}
\author{
Samuel Wieczorek
}
