% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{Get_Code_for_ObserveEvent_widgets}
\alias{Get_Code_for_ObserveEvent_widgets}
\title{Code for declaring xxx}
\usage{
Get_Code_for_ObserveEvent_widgets(widgets.names)
}
\arguments{
\item{widgets.names}{xxx}
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
widgets <- paste0('widget', 1:3)
code <- Get_Code_for_ObserveEvent_widgets(widgets)
cat(code)
}

}
\author{
Samuel Wieczorek
}
