% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetFirstMandatoryNotValidated}
\alias{GetFirstMandatoryNotValidated}
\title{Get the first mandatory step not validated}
\usage{
GetFirstMandatoryNotValidated(range, rv)
}
\arguments{
\item{range}{A `vector` of integers. The min of this vector must be greater
of equal to 0 and the max must be less or equal to the size of the object}

\item{rv}{A `list` with at least two slots :
* mandatory: A vector of `booelan` which indicates whther the steps are
mandatory or not
* steps.status: A vector of `interger()` which indicates the status of the steps}
}
\value{
An integer which is the indice of the identified step in the vector
rv$steps.status
}
\description{
Get the first mandatory step not validated
}
\examples{
NULL

}
