% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetFirstMandatoryNotValidated}
\alias{GetFirstMandatoryNotValidated}
\title{Get First Mandatory Not Validated}
\usage{
GetFirstMandatoryNotValidated(range, steps.info, config)
}
\arguments{
\item{range}{xxx}

\item{steps.info}{xxx}

\item{config}{xxx}
}
\value{
A `integer(1)` which correspond to
}
\description{
xxx
}
