% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetFirstMandatoryNotValidated}
\alias{GetFirstMandatoryNotValidated}
\title{xxx}
\usage{
GetFirstMandatoryNotValidated(range, mandatory, steps.status)
}
\arguments{
\item{range}{xxx}

\item{mandatory}{xxx}

\item{steps.status}{xxx}
}
\value{
xxx
}
\description{
xxx
}
