% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetFirstMandatoryNotValidated}
\alias{GetFirstMandatoryNotValidated}
\title{xxx}
\usage{
GetFirstMandatoryNotValidated(range, rv)
}
\arguments{
\item{range}{xxx}

\item{rv}{xxx}
}
\value{
xxx
}
\description{
xxx
}
