% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{GetHistory}
\alias{GetHistory}
\title{Get the history of an assay}
\usage{
GetHistory(obj, name)
}
\arguments{
\item{obj}{The dataset managed by MagellanNTK}

\item{name}{The name of a slot in the object}
}
\value{
A `data.frame()`
}
\description{
Get the history of an assay
}
\examples{
NULL
}
