% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetListDatasets}
\alias{GetListDatasets}
\title{Get filtered datasets}
\usage{
GetListDatasets(class = NULL, demo_package = NULL)
}
\arguments{
\item{class}{xxx}

\item{demo_package}{xxx}
}
\description{
Get filtered datasets
}
\examples{
foo1 <- GetListDatasets()

}
