% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetListDatasets}
\alias{GetListDatasets}
\title{Get datasets from packages which owns datasets of a given class}
\usage{
GetListDatasets(class = NULL, demo_package = NULL)
}
\arguments{
\item{class}{A `character()` which is the class of the datasets one looking for}

\item{demo_package}{A `character()` which specifies a particular package to search}
}
\value{
A `vector` in which items are the name of datasets which inherits
the class given in parameter
}
\description{
Get datasets from packages which owns datasets of a given class
}
\examples{
foo1 <- GetListDatasets()

}
