% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetListDatasets}
\alias{GetListDatasets}
\title{Get filtered datasets}
\usage{
GetListDatasets(class, filtered = FALSE)
}
\description{
Get filtered datasets
}
\examples{
foo1 <- GetListDatasets()

foo2 <- GetListDatasets(class='QFeatures', filtered = TRUE)


}
