% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{xxx}
\usage{
GetMaxValidated_BeforePos(current.pos, pos = NULL, steps.status)
}
\arguments{
\item{current.pos}{xxx}

\item{pos}{xxx}

\item{steps.status}{xxx}
}
\value{
xxx
}
\description{
xxx
}
