% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_for_pipeline.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{xxx}
\usage{
GetMaxValidated_BeforePos(pos)
}
\arguments{
\item{pos}{xxx}
}
\value{
Nothing
}
\description{
This function calls the server part of each module composing the pipeline
}
