% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{xxx}
\usage{
GetMaxValidated_BeforePos(pos = NULL, rv)
}
\arguments{
\item{pos}{xxx}

\item{rv}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
NULL

}
