% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{Get the last validated step before current position.}
\usage{
GetMaxValidated_BeforePos(pos = NULL, rv)
}
\arguments{
\item{pos}{A `integer(1)` which is the indice of the active position.}

\item{rv}{A `list()` which stores reactiveValues()}
}
\value{
A `integer(1)`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\details{
#' @title xxx
#'
#' @description Encapsulates each UI steps in a <div> tag so as to be able
#' to manage hide/show commands applied on the div itself rether than on the
#' ui.
#'
#' @param ns xxx
#' @param id xxx
#' @param config xxx
#'
#' @return A `renderUI` function
#'
#' @examples NULL
#' 
#' @export
#'
Build_EncapsulateScreens_ui <- function(ns, id, config) {
    len <- length(config@ll.UI)
    if(dev.mode){
      cat ('Entering Build_EncapsulateScreens_ui()')
      show(config)
    }
      
    
    renderUI({
        tagList(
            lapply(seq_len(len), function(i) {
                if (i == 1) {
                    div(
                        id = ns(names(config@steps)[i]),
                        class = paste0("page_", id),
                        config@ll.UI[[i]]
                    )
                } else {
                    shinyjs::hidden(
                        div(
                            id = ns(names(config@steps)[i]),
                            class = paste0("page_", id),
                            config@ll.UI[[i]]
                        )
                    )
                }
            })
        )
    })
}
}
\examples{
NULL

}
