% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetMaxValidated_BeforePos}
\alias{GetMaxValidated_BeforePos}
\title{Get the last validated step before current position.}
\usage{
GetMaxValidated_BeforePos(pos = NULL, rv)
}
\arguments{
\item{pos}{A `integer(1)` which is the indice of the active position.}

\item{rv}{A `list()` which stores reactiveValues()}
}
\value{
A `integer(1)`
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL

}
