% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{GetPackageVersion}
\alias{GetPackageVersion}
\title{Package version
#'description xxx}
\usage{
GetPackageVersion(pkg)
}
\arguments{
\item{pkg}{xxx}
}
\description{
Package version
#'description xxx
}
