% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{GetStringStatus}
\alias{GetStringStatus}
\title{Status to string}
\usage{
GetStringStatus(i, title.style = FALSE)
}
\arguments{
\item{i}{xxx}

\item{title.style}{xxx}
}
\value{
xxx
}
\description{
Converts status code (intefer) into a readable string.
}
