% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{GetValuesFromChildren}
\alias{GetValuesFromChildren}
\title{xxxx}
\usage{
GetValuesFromChildren(config, tmp.return)
}
\arguments{
\item{config}{xxxx}

\item{tmp.return}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
NA

}
