% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{global}
\alias{global}
\title{Describe tags for steps state}
\format{
An object of class \code{list} of length 3.
}
\usage{
global
}
\description{
Describe tags for steps state
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}
}}

\keyword{datasets}
