% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{globalvariables}
\alias{globalvariables}
\alias{port}
\title{Global variables}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
port
}
\description{
Global variables
}
\keyword{datasets}
