% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_history_dataset.R
\name{history_dataset}
\alias{history_dataset}
\alias{history_dataset_ui}
\alias{history_dataset_server}
\title{history_dataset_ui and history_dataset_server}
\usage{
history_dataset_ui(id)

history_dataset_server(
  id,
  dataIn = reactive({
     NULL
 })
)

history_dataset(dataIn)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{dataIn}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(history_dataset(lldata))
}

}
