% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_history_dataset.R
\name{history_dataset}
\alias{history_dataset}
\alias{history_dataset_ui}
\alias{history_dataset_server}
\title{history_dataset_ui and history_dataset_server}
\usage{
history_dataset_ui(id)

history_dataset_server(
  id,
  dataIn = reactive({
     NULL
 }),
  remoteReset = reactive({
     0
 }),
  is.enabled = reactive({
     TRUE
 })
)

history_dataset(obj)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{dataIn}{An instance of the class `MultiAssayExperiment`.}
}
\value{
A shiny app
}
\description{
A shiny Module which show th content of the slot 'history'
in the metadata() of the last SE of the dataset
}
\examples{
if (interactive()){
data(lldata)
shiny::runApp(history_dataset(lldata))
}

}
\keyword{internal}
