% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{infos_dataset}
\alias{infos_dataset}
\alias{infos_dataset_ui}
\alias{infos_dataset_server}
\title{infos_dataset_ui and infos_dataset_server}
\usage{
infos_dataset_ui(id)

infos_dataset_server(
  id,
  dataIn = reactive({
     NULL
 })
)

infos_dataset(dataIn)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(infos_dataset(lldata))
}

}
