% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{infos_dataset}
\alias{infos_dataset}
\alias{infos_dataset_ui}
\alias{infos_dataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
infos_dataset_ui(id)

infos_dataset_server(
  id,
  obj = reactive({
     NULL
 })
)

infos_dataset(obj)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontrun{
  data(lldata)
  shiny::runApp(infos_dataset(lldata))
}

}
