% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{infos_dataset}
\alias{infos_dataset}
\alias{infos_dataset_ui}
\alias{infos_dataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
infos_dataset_ui(id)

infos_dataset_server(
  id,
  obj = reactive({
     NULL
 })
)

infos_dataset(obj)
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
  data(Exp1_R25_prot, package='DaparToolshedData')
  shiny::runApp(infos_dataset(Exp1_R25_prot))

\dontshow{\}) # examplesIf}
}
