% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.substr}
\alias{is.substr}
\title{Substring}
\usage{
is.substr(pattern, target)
}
\arguments{
\item{pattern}{A `character()`}

\item{target}{The `character()` to look for into the pattern}
}
\value{
A boolean
}
\description{
Substring
}
\examples{
is.substr('cde', 'abcdefghi')

}
