% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R
\name{Keep_Items_from_Dataset}
\alias{Keep_Items_from_Dataset}
\title{xxx}
\usage{
Keep_Items_from_Dataset(dataset, range)
}
\arguments{
\item{dataset}{xxx}

\item{range}{xxx}
}
\value{
The dataset minus some items

xxx
}
\description{
Removes one or more items from the dataset. This function is specific of the
type of dataset.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Keep_Items_from_Dataset(Exp1_R25_pept, range = seq_len(2))

}
