% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R, R/mod_navigation.R
\name{Keep_Items_from_Dataset}
\alias{Keep_Items_from_Dataset}
\title{xxx}
\usage{
Keep_Items_from_Dataset(dataset, range)

Keep_Items_from_Dataset(dataset, range)
}
\arguments{
\item{dataset}{xxx}

\item{range}{xxx}
}
\value{
The dataset minus some items

xxx

The dataset minus some items

xxx
}
\description{
Removes one or more items from the dataset. This function is specific of the
type of dataset.

Removes one or more items from the dataset. This function is specific of the
type of dataset.
}
\examples{
obj <- Keep_Items_from_Dataset(QFeatures::feat1, range = seq_len(1))

obj <- Keep_Items_from_Dataset(QFeatures::feat1, range = seq_len(1))

}
