% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Load_Dataset.R
\name{Load_Dataset}
\alias{Load_Dataset}
\alias{Load_Dataset_ui}
\alias{Load_Dataset_server}
\title{Load dataset shiny module}
\usage{
Load_Dataset_ui(id)

Load_Dataset_server(id)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
