% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_workflow.R
\name{Load_Dataset}
\alias{Load_Dataset}
\alias{mod_load_workflow_ui}
\alias{mod_load_workflow_server}
\title{Load dataset shiny module}
\usage{
mod_load_workflow_ui(id)

mod_load_workflow_server(
  id,
  path = reactive({
     NULL
 }),
  usermod = "user"
)
}
\arguments{
\item{id}{xxx}

\item{path}{xxx}

\item{mode}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
