% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_workflow.R
\name{Load_Dataset}
\alias{Load_Dataset}
\alias{mod_load_workflow_ui}
\alias{mod_load_workflow_server}
\title{Load dataset shiny module}
\usage{
mod_load_workflow_ui(id)

mod_load_workflow_server(
  id,
  path = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
\examples{

if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    Load_Dataset_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    Load_Dataset_server("demo")
  }
  
  shinyApp(ui, server)
}
}
