% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_workflow.R, R/Load_Dataset_modal.R,
%   R/Load_Dataset.R
\name{Load_Dataset}
\alias{Load_Dataset}
\alias{Load_Workflow_ui}
\alias{Load_Workflow_server}
\alias{Load_Dataset_modal_ui}
\alias{Load_Dataset_modal_server}
\alias{Load_Dataset_ui}
\alias{Load_Dataset_server}
\title{Load dataset shiny module}
\usage{
Load_Workflow_ui(id)

Load_Workflow_server(id)

Load_Dataset_modal_ui(id)

Load_Dataset_modal_server(id)

Load_Dataset_ui(id)

Load_Dataset_server(
  id,
  path = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx

xxxxx

xxxxx
}
\description{
A shiny Module to load a dataset.

A shiny Module to load a dataset.

A shiny Module to load a dataset.
}
\examples{

if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    Load_Dataset_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    Load_Dataset_server("demo", path = reactive({NULL}))
  }
  
  shinyApp(ui, server)
}

if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    Load_Dataset_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    Load_Dataset_server("demo", path = reactive({NULL}))
  }
  
  shinyApp(ui, server)
}

if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    Load_Dataset_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    Load_Dataset_server("demo", path = reactive({NULL}))
  }
  
  shinyApp(ui, server)
}
}
