% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanNTK.R
\name{magellanNTK}
\alias{magellanNTK}
\alias{MagellanNTK_ui}
\alias{MagellanNTK_server}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK_ui(id, sidebarSize = "medium")

MagellanNTK_server(
  id,
  dataIn = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  workflow.name = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "user"
)

MagellanNTK(
  dataIn = NULL,
  workflow.path = NULL,
  workflow.name = NULL,
  convert.path = NULL,
  verbose = FALSE,
  usermod = "user",
  sidebarSize = "medium",
  ...
)
}
\arguments{
\item{id}{xxx}

\item{sidebarSize}{The width of the sidebar. Available values are 'small', 'medium', 'large'}

\item{dataIn}{A `list()` of data.frame like items. This parameter must be
used with '[[]]' to access items}

\item{workflow.path}{xxx}

\item{workflow.name}{xxxx}

\item{verbose}{A `boolean(1)`}

\item{usermod}{xxx}

\item{convert.path}{xxx}

\item{...}{xxx}
}
\value{
NA
}
\description{
xxx

DO NOT REMOVE.
}
\details{
The list of customizable funcs (param `funcs`) contains the following items:

 These are the default values where each item points to a default fucntion
 implemented into MagellanNTK.

 The user can modify these values by two means:
 * setting the values in the parameter to pass to the function
 `MagellanNTK()`,
 * inside the UI of MagellanNTK, in the settings panels
}
\examples{
if (interactive()) {
    MagellanNTK()
}

if (interactive()) {
    # launch without initial config
    shiny::runApp(MagellanNTK())
}

}
