% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanNTK.R
\name{MagellanNTK}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK(
  obj = NULL,
  workflow.path = NULL,
  workflow.name = NULL,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{A `boolean(1)`}

\item{funcs}{An list containing th following items:
* title: xxx
* base_URL: xxx}
}
\description{
xxx
}
\details{
The list of customizable funcs (param `funcs`) contains the following items:

 These are the default values where each item points to a default fucntion
 implemented into MagellanNTK.
 
 The user can modify these values by two means:
 * setting the values in the parameter to pass to the function 
 `MagellanNTK()`,
 * inside the UI of MagellanNTK, in the settings panels
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# launch without initial config
shiny::runApp(MagellanNTK())


MagellanNTK(funcs)
\dontshow{\}) # examplesIf}
}
