% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MagellanNTK.R
\name{magellanNTK}
\alias{magellanNTK}
\alias{MagellanNTK_ui}
\alias{MagellanNTK_server}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK_ui(id)

MagellanNTK_server(
  id,
  obj = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  workflow.name = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "dev"
)

MagellanNTK(
  obj = NULL,
  workflow.path = NULL,
  workflow.name = NULL,
  verbose = FALSE,
  usermod = "dev"
)
}
\arguments{
\item{obj}{xxx}

\item{workflow.path}{xxxx}

\item{verbose}{A `boolean(1)`}

\item{usermode}{xxx}

\item{request}{Internal parameter for `{shiny}`. 
DO NOT REMOVE.}

\item{input, output, session}{Internal parameters for {shiny}. 
DO NOT REMOVE.}
}
\description{
xxx
}
\details{
The list of customizable funcs (param `funcs`) contains the following items:

 These are the default values where each item points to a default fucntion
 implemented into MagellanNTK.
 
 The user can modify these values by two means:
 * setting the values in the parameter to pass to the function 
 `MagellanNTK()`,
 * inside the UI of MagellanNTK, in the settings panels
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
MagellanNTK()
\dontshow{\}) # examplesIf}
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# launch without initial config
shiny::runApp(MagellanNTK())
\dontshow{\}) # examplesIf}
}
