% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MagellanNTK}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK(funcs = default.funcs, verbose = FALSE)
}
\arguments{
\item{funcs}{An list containing th following items:
* title: xxx
* base_URL: xxx}

\item{workflow}{A list containing the following items:
* name : The name of the workflow. If it is in a package, the functions
name_ui(), name_server() and name_config() must be be available in its 
environment
* xxx}
}
\description{
xxx
}
\details{
The list of customizable funcs (param `funcs`) contains the following items:

 These are the default values where each item points to a default fucntion
 implemented into MagellanNTK.
 
 The user can modify these values by two means:
 * setting the values in the parameter to pass to the function 
 `MagellanNTK()`,
 * inside the UI of MagellanNTK, in the settings panels
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# launch without initial config
shiny::runApp(MagellanNTK())

# launch with config
funcs <- list(convert_dataset = "DaparToolshed::convert_dataset",
open_dataset = "DaparToolshed::open_dataset",
open_demoDataset = "DaparToolshed::open_demoDataset",
view_dataset = "omXplore::view_dataset",
infos_dataset = "DaparToolshed::infos_dataset",
addDatasets = "DaparToolshed::addDatasets",
keepDatasets = "DaparToolshed::keepDatasets" )

funcs <- list(convert_dataset = "MagellanNTK::convert",
  open_dataset = "MagellanNTK::open_dataset",
  open_demoDataset = "MagellanNTK::open_demoDataset",
  view_dataset = "MagellanNTK::view_dataset",
  infos_dataset = "MagellanNTK::infos_dataset",
  addDatasets = "MagellanNTK::addDatasets",
  keepDatasets = "MagellanNTK::keepDatasets")

MagellanNTK(funcs)

\dontshow{\}) # examplesIf}
}
