% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MagellanNTK}
\alias{MagellanNTK}
\title{Main Shiny application}
\usage{
MagellanNTK(config)
}
\description{
xxx
}
\examples{
if(interactive()){

# launch without initial config
MagellanNTK()

# launch with config
funcs <- list(convert = "DaparToolshed::convert",
open_dataset = "DaparToolshed::open_dataset",
open_demoDataset = "DaparToolshed::open_demoDataset",
view_dataset = "DaparViz::view_dataset",
infos_dataset = "DaparToolshed::infos_dataset")

funcs <- list(convert = "MagellanNTK::convert",
  open_dataset = "MagellanNTK::open_dataset",
  open_demoDataset = "MagellanNTK::open_demoDataset",
  view_dataset = "MagellanNTK::view_dataset",
  infos_dataset = "MagellanNTK::infos_dataset")

MagellanNTK(funcs)
}

}
