% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mdEditor.R
\name{mdEditor}
\alias{mdEditor}
\alias{mdEditor_ui}
\alias{mdEditor_server}
\title{Markdown editor module}
\usage{
mdEditor_ui(id)

mdEditor_server(id)

mdEditor()
}
\arguments{
\item{id}{A `character(1)` which is the id of the instance of the module}
}
\value{
A Shiny modal-dialog

A Shiny modal-dialog
}
\description{
Displays of formatted modal-dialog with 'Cancel' and 
'Ok' buttons.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mdEditor())

\dontshow{\}) # examplesIf}
}
