% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bsmodal.R
\name{mod_bsmodal_server}
\alias{mod_bsmodal_server}
\title{xxxx}
\usage{
mod_bsmodal_server(id, title = NULL, width = NULL, uiContent = NULL)
}
\arguments{
\item{id}{xxxx}

\item{title}{xxx}

\item{width}{xxx}

\item{uiContent}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\examples{
\donttest{ 
library(shiny)
library(shinyBS)

ui <- fluidPage(
  mod_bsmodal_ui('tbl')
)
server <- function(input, output){
  mod_bsmodal_server(id = 'tbl',
                     title = 'test',
                     uiContent = p('test'))
}
shinyApp(ui, server)
}

}
