% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Debug_Infos.R
\name{mod_Debug_Infos_ui}
\alias{mod_Debug_Infos_ui}
\alias{mod_Debug_Infos_server}
\title{mod_Debug_Infos}
\usage{
mod_Debug_Infos_ui(id)

mod_Debug_Infos_server(
  id,
  title = NULL,
  config = reactive({
     NULL
 }),
  rv.dataIn = reactive({
     NULL
 }),
  dataIn = reactive({
     NULL
 }),
  dataOut = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  steps.status = reactive({
     NULL
 }),
  steps.skipped = reactive({
     NULL
 }),
  current.pos = reactive({
     NULL
 }),
  is.enabled = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{xxx}

\item{title}{xxx}

\item{config}{A `list`}

\item{rv.dataIn}{xxx}

\item{dataIn}{An instance of the class `QFeatures`}

\item{dataOut}{An instance of the class `QFeatures`}

\item{steps.enabled}{A `logical()` xxxx}

\item{steps.status}{A `logical()` xxxx}

\item{steps.skipped}{A `logical()` xxxx}

\item{current.pos}{A `integer(1)` xxxx}

\item{is.enabled}{A `logical(1)` xxxx}
}
\value{
NA
}
\description{
A shiny Module.

xxx
}
\examples{
if (interactive()) {
    ui <- fluidPage(
        mod_format_DT_ui("tbl")
    )
    server <- function(input, output) {
        mod_format_DT_server(id = "tbl", table2show = reactive({
            head(iris)
        }))
    }
    shinyApp(ui, server)
}

}
