% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Debug_Infos.R
\name{mod_Debug_Infos_ui}
\alias{mod_Debug_Infos_ui}
\alias{mod_Debug_Infos_server}
\title{mod_Debug_Infos}
\usage{
mod_Debug_Infos_ui(id)

mod_Debug_Infos_server(
  id,
  title = NULL,
  config = reactive({     NULL }),
  rv.dataIn = reactive({     NULL }),
  dataIn = reactive({     NULL }),
  dataOut = reactive({     NULL }),
  steps.infos = reactive({     NULL }),
  current.pos = reactive({     NULL }),
  is.enabled = reactive({     NULL })
)
}
\arguments{
\item{id}{xxx}

\item{title}{xxx}

\item{config}{xxx}

\item{rv.dataIn}{xxx}

\item{dataIn}{xxx}

\item{dataOut}{xxxx}

\item{current.pos}{xxx}

\item{is.enabled}{xxx}

\item{steps.status}{xxx}

\item{steps.enabled}{xxxx}
}
\value{
xxx
}
\description{
A shiny Module.

xxx
}
\examples{
\dontrun{

ui <- fluidPage(
  mod_format_DT_ui('tbl')
)
server <- function(input, output){
  mod_format_DT_server(id = 'tbl',table2show = reactive({head(iris)}))
}
shinyApp(ui, server)
}

}
