% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_btns.R
\name{mod_download_btns}
\alias{mod_download_btns}
\alias{mod_download_btns_ui}
\alias{mod_download_btns_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_download_btns_ui(id)

mod_download_btns_server(id, df.data, name, style = NULL)
}
\arguments{
\item{id}{internal}

\item{df.data}{internal}

\item{name}{internal}

\item{style}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\examples{
if(interactive()){
  
  
  
}
}
\keyword{internal}
