% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_errorModal.R
\name{mod_err_modal}
\alias{mod_err_modal}
\alias{mod_errorModal_ui}
\alias{mod_errorModal_server}
\alias{mod_errorModal}
\title{A modal window to display messages}
\usage{
mod_errorModal_ui(id)

mod_errorModal_server(
  id,
  title = NULL,
  text = NULL,
  footer = modalButton("Close")
)

mod_errorModal(title = NULL, text = NULL)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{title}{A `character()` as the title of the modal}

\item{text}{A `character()` as the content of the modal}

\item{footer}{The content of the footer . May be UI content}
}
\value{
A shiny App
}
\description{
This module is not directly used by MagellanNTK core functions.
It is rather a useful tool for third part pipelines and processes.
}
\examples{
if (interactive()) {
    shiny::runApp(mod_errorModal("myTitle", "myContent"))
}

}
