% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_errorModal.R
\name{mod_err_modal}
\alias{mod_err_modal}
\alias{mod_errorModal_ui}
\alias{mod_errorModal_server}
\alias{mod_errorModal}
\title{xxx}
\usage{
mod_errorModal_ui(id)

mod_errorModal_server(
  id,
  title = NULL,
  text = NULL,
  footer = modalButton("Close")
)

mod_errorModal(title = NULL, text = NULL)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{title}{xxxx}

\item{text}{xxx}

\item{footer}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
if (interactive()) {
    shiny::runApp(mod_errorModal("myTitle", "myContent"))
}

}
