% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT_server}
\alias{mod_format_DT_server}
\title{xxx}
\usage{
mod_format_DT_server(
  id,
  table2show,
  withBtns = NULL,
  showRownames = FALSE,
  dom = "Bt",
  style = reactive({     NULL })
)
}
\arguments{
\item{withBtns}{xxx}

\item{showRownames}{xxxx}

\item{dom}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\value{
xxx
}
\description{
xxx
}
\examples{
library(shiny)
library(DT)

ui <- fluidPage(
  mod_format_DT_ui('tbl')
)
server <- function(input, output){
  mod_format_DT_server(id = 'tbl',table2show = reactive({head(iris)}))
}
shinyApp(ui, server)

}
\keyword{internal}
