% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_insert_md.R
\name{mod_insert_md}
\alias{mod_insert_md}
\alias{insert_md_ui}
\alias{insert_md_server}
\alias{insert_md}
\title{insert_md_ui and insert_md_server}
\usage{
insert_md_ui(id)

insert_md_server(id, url, link_URL = NULL)

insert_md(url)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{url}{internal}

\item{link_URL}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    base <- system.file("app/md", package = "MagellanNTK")
    url <- file.path(base, "presentation.Rmd")
    shiny::runApp(insert_md(url))
}
}
