% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_insert_md.R
\name{mod_insert_md}
\alias{mod_insert_md}
\alias{insert_md_ui}
\alias{insert_md_server}
\alias{insert_md}
\title{mod_insert_md_ui and mod_insert_md_server}
\usage{
insert_md_ui(id)

insert_md_server(id, url, link_URL = NULL)

insert_md(url)
}
\arguments{
\item{id}{shiny id}

\item{url}{internal}

\item{link_URL}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
base <- system.file('app/md', package = 'MagellanNTK')
url <- file.path(base, "presentation.Rmd")
shiny::runApp(insert_md(url))

\dontshow{\}) # examplesIf}
}
