% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Load_Dataset.R
\name{mod_Load_Dataset}
\alias{mod_Load_Dataset}
\alias{mod_Load_Dataset_ui}
\alias{mod_Load_Dataset_server}
\title{Load dataset shiny module}
\usage{
mod_Load_Dataset_ui(id)

mod_Load_Dataset_server(id)
}
\arguments{
\item{id}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
\examples{

if (interactive()){

  library(shiny)
  options(shiny.fullstacktrace = TRUE)
  
  
  ui <- fluidPage(
    mod_Load_Dataset_ui("demo")
  )
  
  server <- function(input, output, session) {
    data(data_na)
    mod_Load_Dataset_server("demo")
  }
  
  shinyApp(ui, server)
}
}
