% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Load_Dataset.R
\name{mod_Load_Dataset}
\alias{mod_Load_Dataset}
\alias{mod_Load_Dataset_ui}
\alias{mod_Load_Dataset_server}
\title{Load dataset shiny module}
\usage{
mod_Load_Dataset_ui(id)

mod_Load_Dataset_server(id)
}
\arguments{
\item{id}{xxx}
}
\value{
NA
}
\description{
A shiny Module to load a dataset.
}
\examples{
if (interactive()) {
    data(ft_na)
    ui <- mod_Load_Dataset_ui("load")

    server <- function(input, output, session) {
        mod_Load_Dataset_server(id = "load")}
    shinyApp(ui = ui, server = server)
}
}
