% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_load_package.R
\name{mod_load_package}
\alias{mod_load_package}
\alias{mod_load_package_ui}
\alias{mod_load_package_server}
\alias{load_package}
\title{Change the default functions in `MagellanNTK`}
\usage{
mod_load_package_ui(id)

mod_load_package_server(
  id,
  funcs = reactive({
     NULL
 })
)

load_package(funcs = NULL)
}
\arguments{
\item{id}{shiny id}

\item{funcs}{A list}
}
\description{
This module allows to change the default functions
embedded in the package `MagellanNTK`. These fucntions are the following:
* convert_dataset: xxx
* view_dataset: xxx
* infos_dataset: xxx
}
\examples{
\dontrun{

funcs <- list(convert_dataset = "DaparToolshed::convert_dataset",
open_dataset = "MagellanNTK::open_dataset",
open_demoDataset = "MagellanNTK::open_demoDataset",
view_dataset = "omXplore::view_dataset",
infos_dataset = "MagellanNTK::infos_dataset",
download_dataset = "MagellanNTK::download_dataset",
export_dataset = "MagellanNTK::export_dataset",
addDatasets = "MagellanNTK::addDatasets",
keepDatasets = "MagellanNTK::keepDatasets"
)
shiny::runApp(load_package(funcs))

shiny::runApp(load_package())
}

}
\author{
Samuel Wieczorek
}
