% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainapp.R
\name{mod_main_page}
\alias{mod_main_page}
\alias{mainapp_ui}
\alias{mainapp_server}
\alias{mainapp}
\title{mod_main_page_ui and mod_loading_page_server}
\usage{
mainapp_ui(id, session)

mainapp_server(
  id,
  dataIn = reactive({
     NULL
 }),
  workflow.name = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "dev"
)

mainapp(usermod = "dev")
}
\arguments{
\item{id}{shiny id}

\item{session}{xxx}

\item{dataIn}{xxx}

\item{workflow.name}{= reactive({NULL}),}

\item{workflow.path}{= reactive({NULL}),}

\item{verbose}{= FALSE,}

\item{usermod}{= 'dev'}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
shiny::runApp(mainapp())
}

}
