% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainapp.R
\name{mod_main_page}
\alias{mod_main_page}
\alias{mainapp_ui}
\alias{mainapp_server}
\alias{mainapp}
\title{mod_main_page_ui and mod_loading_page_server}
\usage{
mainapp_ui(id, session, size = "300px")

mainapp_server(
  id,
  dataIn = reactive({
     NULL
 }),
  data.name = reactive({
     "myDataset"
 }),
  workflow.name = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  verbose = FALSE,
  usermod = "user"
)

mainapp(usermod = "user")
}
\arguments{
\item{id}{shiny id}

\item{session}{xxx}

\item{size}{The width of the sidebar.in pixels}

\item{dataIn}{xxx}

\item{data.name}{The name of the dataset. Default is 'myDataset'}

\item{workflow.name}{Default is NULL,}

\item{workflow.path}{Default is NULL,}

\item{verbose}{= FALSE,}

\item{usermod}{= 'dev'}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(mainapp())
}

}
