% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mainapp.R
\name{mod_main_page}
\alias{mod_main_page}
\alias{mainapp_ui}
\alias{mainapp_server}
\alias{mainapp}
\title{mod_main_page_ui and mod_loading_page_server}
\usage{
mainapp_ui(id)

mainapp_server(
  id,
  obj = reactive({
     NULL
 }),
  workflow.name = reactive({
     NULL
 }),
  workflow.path = reactive({
     NULL
 }),
  verbose = FALSE
)

mainapp()
}
\arguments{
\item{id}{shiny id}

\item{verbose}{xxx}

\item{funcs}{xxx}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mainapp())

\dontshow{\}) # examplesIf}
}
