% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nav_pipeline.R
\name{mod_nav_pipeline_server}
\alias{mod_nav_pipeline_server}
\title{xxx}
\usage{
mod_nav_pipeline_server(
  id,
  dataIn = reactive({     NULL }),
  is.enabled = reactive({     TRUE }),
  remoteReset = reactive({     FALSE }),
  is.skipped = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{is.enabled}{xxx}

\item{remoteReset}{xxx}

\item{is.skipped}{xxx}
}
\value{
Nothing

Nothing

Nothing.

Nothing
}
\description{
xxxxxx

xxx

This function calls the server part of each module composing the pipeline

Catch the return value of a module and update the list of isDone modules
This list is updated with the names of datasets present in the rv$tmp
variable. One set to TRUE all the elements in isDone which have a corresponding
element in names(rv$tmp).
One cannot simply set to TRUE the last element of rv$tmp because it will does
not work in case of a reseted module (it is not in the names(rv$tmp) list
anymore)
If a value (not NULL) is received, then it corresponds to the module
pointed by the current position
This function also updates the list isDone
This function updates the current dataset (self$rv$dataIn)

et to skipped all steps of the current object

xxx
}
\section{Fields}{

\describe{
\item{\code{modal_txt}}{xxx}
}}

