% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{mod_open_dataset}
\alias{mod_open_dataset}
\alias{open_dataset_ui}
\alias{open_dataset_server}
\alias{open_dataset}
\title{mod_open_dataset_ui and mod_open_dataset_server}
\usage{
open_dataset_ui(id)

open_dataset_server(id)

open_dataset()
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
shiny::runApp(open_dataset())
}

}
\keyword{internal}
