% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_workflow.R
\name{mod_open_workflow}
\alias{mod_open_workflow}
\alias{open_workflow_ui}
\alias{open_workflow_server}
\alias{open_workflow}
\title{mod_open_workflow_ui and mod_open_workflow_server}
\usage{
open_workflow_ui(id)

open_workflow_server(id)

open_workflow()
}
\arguments{
\item{id}{xxx}
}
\value{
A list
}
\description{
A shiny Module.
}
\examples{
\dontrun{
shiny::runApp(open_workflow())
}

}
