% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popover_for_help.R
\name{mod_popover_for_help}
\alias{mod_popover_for_help}
\alias{mod_popover_for_help_ui}
\alias{mod_popover_for_help_server}
\alias{popover_for_help}
\title{mod_popover_for_help_ui and mod_popover_for_help_server}
\usage{
mod_popover_for_help_ui(id)

mod_popover_for_help_server(id, title, content)

popover_for_help(title, content)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{title}{xxxx}

\item{content}{xxx}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    shiny::runApp(popover_for_help("myTitle", "myContent"))
}

}
