% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popover_for_help.R
\name{mod_popover_for_help_ui}
\alias{mod_popover_for_help_ui}
\alias{mod_popover_for_help_server}
\title{mod_popover_for_help_ui and mod_popover_for_help_server}
\usage{
mod_popover_for_help_ui(id)

mod_popover_for_help_server(id, data)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}
}
\value{
xxx
}
\description{
A shiny Module.

xxx
}
\examples{
\donttest{
library(shiny)
library(shinyBS)
ui <- fluidPage(
  mod_popover_for_help_ui('Title')
)
server <- function(input, output){
  mod_popover_for_help_server(id = 'Title',
                              data = list(title = 'Test', 
                                          content = 'Test')
  )
}
shinyApp(ui, server)
}

}
