% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popover_for_help.R
\name{mod_popover_for_help_server}
\alias{mod_popover_for_help_server}
\title{xxx}
\usage{
mod_popover_for_help_server(id, data)
}
\value{
xxx
}
\description{
xxx
}
\examples{
library(shiny)
library(shinyBS)
ui <- fluidPage(
  mod_popover_for_help_ui('Title')
)
server <- function(input, output){
  mod_popover_for_help_server(id = 'Title',
                              data = list(title = 'Test', 
                                          content = 'Test')
  )
}
shinyApp(ui, server)


}
\keyword{internal}
