% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_release_notes.R
\name{mod_release_notes}
\alias{mod_release_notes}
\alias{mod_release_notes_ui}
\alias{mod_release_notes_server}
\alias{release_notes}
\title{mod_release_notes_ui and mod_release_notes_server}
\usage{
mod_release_notes_ui(id)

mod_release_notes_server(id, URL_releaseNotes)

release_notes(URL_releaseNotes)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{URL_releaseNotes}{The path to the Rmd file to display. It can be a path
to a file on the computer or a link to a file over internet.}
}
\value{
A shiny App
}
\description{
A shiny Module.
}
\examples{
if (interactive()) {
    url <- "http://www.prostar-proteomics.org/md/versionNotes.md"
    shiny::runApp(release_notes(url))
}

}
