% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_release_notes.R
\name{mod_release_notes_ui}
\alias{mod_release_notes_ui}
\alias{mod_release_notes_server}
\alias{release_notes}
\title{mod_release_notes_ui and mod_release_notes_server}
\usage{
mod_release_notes_ui(id)

mod_release_notes_server(id)

release_notes()
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
