% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Save_dataset.R
\name{mod_Save_Dataset}
\alias{mod_Save_Dataset}
\alias{mod_Save_Dataset_ui}
\alias{mod_Save_Dataset_server}
\title{Load dataset shiny module}
\usage{
mod_Save_Dataset_ui(id)

mod_Save_Dataset_server(id, data)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
