% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Save_Dataset.R
\name{mod_Save_Dataset_ui}
\alias{mod_Save_Dataset_ui}
\alias{mod_Save_Dataset_server}
\title{mod_Load_Dataset}
\usage{
mod_Save_Dataset_ui(id)

mod_Save_Dataset_server(id, dataIn = reactive({
    NULL
}))
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
