% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sweetAlert.R
\name{mod_sweetAlert}
\alias{mod_sweetAlert}
\alias{mod_SweetAlert_ui}
\alias{mod_SweetAlert_server}
\alias{mod_SweetAlert}
\title{xxx}
\usage{
mod_SweetAlert_ui(id)

mod_SweetAlert_server(
  id,
  title = NULL,
  text = NULL,
  showClipBtn = TRUE,
  type = "error"
)

mod_SweetAlert(title, text, type = "warning")
}
\arguments{
\item{id}{xxx}

\item{title}{xxx}

\item{text}{xxx}

\item{showClipBtn}{xxx}

\item{type}{xxx}
}
\description{
xxxxx
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(mod_SweetAlert('my title', 'my message'))
\dontshow{\}) # examplesIf}
}
