% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timeline_v.R
\name{mod_timeline_v_server}
\alias{mod_timeline_v_server}
\title{timeline Server Function}
\usage{
mod_timeline_v_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{xxx}

\item{config}{xxx}

\item{status}{xxx}

\item{position}{xxx}

\item{enabled}{xxx}
}
\value{
xxx
}
\description{
timeline Server Function
}
\examples{
ui <- fluidPage(
actionButton('changePos', 'Change position'),
mod_timeline_v_ui('TLh')
)

server <- function(input, output){
 rv <- reactiveValues(
 status = c(0, 1, 0, 0),
 current.pos = 1,
 tl.tags.enabled = c(1, 1, 1, 1),
 position = NULL
 )
 config <- list(name = 'Protein_Filtering',
 steps = c('Description', 'Step1', 'Step2', 'Step3'),
 mandatory = c(TRUE, FALSE, TRUE, TRUE)
 )
 mod_timeline_v_server(id = 'TLh',
  config = config,
   status = reactive({rv$status}),
   position = reactive({rv$current.pos}),
   enabled = reactive({rv$tl.tags.enabled})
   )
   }
   shinyApp(ui, server)
}
