% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_used_by_modules.R
\name{Module_Return_Func}
\alias{Module_Return_Func}
\title{Code for declaring xxx}
\usage{
Module_Return_Func()
}
\description{
This function xxx
# Generate dynamically the observeEvent function for each widget
}
\examples{
\dontrun{
code <- Module_Return_Func()
cat(code)
}

}
\author{
Samuel Wieczorek
}
