% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_pipeline.R
\name{nav_pipeline}
\alias{nav_pipeline}
\alias{nav_pipeline_ui}
\alias{nav_pipeline_server}
\title{The server() function of the module `nav_pipeline`}
\usage{
nav_pipeline_ui(id)

nav_pipeline_server(
  id = NULL,
  dataIn = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 }),
  remoteReset = reactive({
     0
 }),
  verbose = FALSE,
  usermod = "user"
)

nav_pipeline()
}
\arguments{
\item{id}{A `character(1)` which defines the id of the module.
It is the same as for the server() function.}

\item{dataIn}{An instance of the `QFeatures` class}

\item{is.enabled}{A `boolean`. This variable is a remote command to specify
if the corresponding module is enabled/disabled in the calling module of
upper level.
For example, if this module is part of a pipeline and the pipeline calculates
that it is disabled (i.e. skipped), then this variable is set to TRUE. Then,
all the widgets will be disabled. If not, the enabling/disabling of widgets
is deciding by this module.}

\item{remoteReset}{An `integer` which acts as a remote command to reset the 
module. Its value is incremented on a external event and it is used to 
trigger an event 
in this module}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}

\item{usermod}{A character to specifies the running mode of MagellanNTK. 
* user (default) : xxx
* dev: xxx}
}
\value{
A list of four items:
* dataOut A dataset of the same class of the parameter dataIn
* steps.enabled A vector of `boolean` of the same length than config@steps
* status A vector of `integer(1)` of the same length than the config@steps
  vector
}
\description{
The module navigation can be launched via a Shiny app.
This is the core module of MagellanNTK
}
\examples{
if (interactive()) {
    library(shiny)
    nav_pipeline()
}

}
\author{
Samuel Wieczorek
}
