% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_process.R
\name{nav_process}
\alias{nav_process}
\alias{nav_process_ui}
\alias{nav_process_server}
\title{The server() function of the module `nav_process`}
\usage{
nav_process_ui(id)

nav_process_server(
  id = NULL,
  dataIn = reactive({
     NULL
 }),
  status = reactive({
     NULL
 }),
  is.enabled = reactive({
     TRUE
 }),
  remoteReset = reactive({
     0
 }),
  remoteResetUI = reactive({
     0
 }),
  is.skipped = reactive({
     FALSE
 }),
  verbose = FALSE,
  usermod = "user",
  btnEvents = reactive({
     NULL
 }),
  runmode = NULL
)

nav_process()
}
\arguments{
\item{id}{A `character(1)` which defines the id of the module.
It is the same as for the server() function.}

\item{dataIn}{The dataset}

\item{status}{xxx}

\item{is.enabled}{A `boolean`. This variable is a remote command to specify
if the corresponding module is enabled/disabled in the calling module of
upper level.
For example, if this module is part of a pipeline and the pipeline calculates
that it is disabled (i.e. skipped), then this variable is set to TRUE. Then,
all the widgets will be disabled. If not, the enabling/disabling of widgets
is deciding by this module.}

\item{remoteReset}{It is a remote command to reset the module. A boolen that
indicates is the pipeline has been reseted by a program of higher level
Basically, it is the program which has called this module}

\item{remoteResetUI}{xxx}

\item{is.skipped}{xxx}

\item{verbose}{= FALSE,}

\item{usermod}{= 'user'}

\item{btnEvents}{xxxx}

\item{runmode}{xxx}
}
\value{
A list of four items:
* dataOut A dataset of the same class of the parameter dataIn
* steps.enabled A vector of `boolean` of the same length than config@steps
* status A vector of `integer(1)` of the same length than the config@steps
  vector
* reset xxxx
}
\description{
The module navigation can be launched via a Shiny app.
This is the core module of MagellanNTK
}
\examples{
if (interactive()) {
    nav_process()
}

}
\author{
Samuel Wieczorek
}
