% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_singleProcess.R
\name{nav_single_process}
\alias{nav_single_process}
\alias{nav_single_process_ui}
\alias{nav_single_process_server}
\title{Shiny module `nav_single_process()`}
\usage{
nav_single_process_ui(id)

nav_single_process_server(
  id = NULL,
  dataIn = reactive({
     NULL
 }),
  status = reactive({
     NULL
 }),
  remoteResetUI = reactive({
     0
 }),
  is.enabled = TRUE,
  is.skipped = FALSE,
  verbose = FALSE,
  usermod = "user",
  btnEvents = reactive({
     NULL
 })
)

nav_single_process()
}
\arguments{
\item{id}{A `character(1)` which defines the id of the module.
It is the same as for the server() function.}

\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{status}{xxx}

\item{remoteResetUI}{xxx}

\item{is.enabled}{xxxx}

\item{is.skipped}{is.skipped xxx}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}

\item{usermod}{= 'user'}

\item{btnEvents}{xxxx}

\item{history}{xxx}
}
\value{
A list of four items:
* dataOut A dataset of the same class of the parameter dataIn
* steps.enabled A vector of `boolean` of the same length than config@steps
* status A vector of `integer(1)` of the same length than the config@steps
  vector
* reset xxxx
}
\description{
The module navigation can be launched via a Shiny app.
This is the core module of MagellanNTK
}
\examples{
if (interactive()) {
    nav_single_process()
}

}
\author{
Samuel Wieczorek
}
