% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{NavPage}
\alias{NavPage}
\title{xxx}
\usage{
NavPage(direction, current.pos, len)
}
\arguments{
\item{direction}{A `integer(1)` which is the direction of the xxx:
forward ('1'), backwards ('-1').}

\item{current.pos}{A `integer(1)` which is the current position.}

\item{len}{A `integer(1)` which is the number of steps in the process.}
}
\value{
A `integer(1)` which is the new current position.
}
\description{
xxx
}
