% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{NavPage}
\alias{NavPage}
\title{xxx}
\usage{
NavPage(direction, current.pos, len)
}
\arguments{
\item{direction}{xxx}

\item{current.pos}{xxx}

\item{len}{xxx}
}
\value{
xxx
}
\description{
xxx
}
