% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{optionsBtnClass}
\alias{optionsBtnClass}
\title{xxxx}
\format{
An object of class \code{character} of length 1.
}
\usage{
optionsBtnClass
}
\description{
xxx
}
\section{Fields}{

\describe{
\item{\code{btn_style}}{xxxx}
}}

\keyword{datasets}
