% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{PrepareData2Send}
\alias{PrepareData2Send}
\title{xxxx}
\usage{
PrepareData2Send(
  steps.info,
  pos,
  temp.dataIn,
  dataIn,
  config,
  current.pos,
  original.length
)
}
\arguments{
\item{pos}{xxx}

\item{rv}{xxxx}
}
\description{
xxx
}
