% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_funcs.R
\name{PrepareData2Send}
\alias{PrepareData2Send}
\title{Prepare Data to Send}
\usage{
PrepareData2Send(
  steps.info,
  temp.dataIn,
  dataIn,
  config,
  current.pos,
  original.length
)
}
\arguments{
\item{steps.info}{xxx}

\item{temp.dataIn}{xxx}

\item{dataIn}{xxx}

\item{config}{xxx}

\item{current.pos}{xxx}

\item{original.length}{xxx}
}
\description{
xxx
}
