% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Process.R
\name{Process}
\alias{Process}
\title{Add together two numbers}
\description{
Add together two numbers

Add together two numbers
}
\examples{

## ------------------------------------------------
## Method `Process$ToggleState_Screens`
## ------------------------------------------------

add(1, 1)
}
\section{Super class}{
\code{\link[Magellan:ScreenManager]{Magellan::ScreenManager}} -> \code{Process}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ToggleState_Screens}{\code{Process$ToggleState_Screens()}}
\item \href{#method-Set_All_Skipped}{\code{Process$Set_All_Skipped()}}
\item \href{#method-Discover_Skipped_Steps}{\code{Process$Discover_Skipped_Steps()}}
\item \href{#method-Set_All_Reset}{\code{Process$Set_All_Reset()}}
\item \href{#method-ValidateCurrentPos}{\code{Process$ValidateCurrentPos()}}
\item \href{#method-EncapsulateScreens}{\code{Process$EncapsulateScreens()}}
\item \href{#method-GetScreens_ui}{\code{Process$GetScreens_ui()}}
\item \href{#method-clone}{\code{Process$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ActionOn_NewPosition">}\href{../../Magellan/html/ScreenManager.html#method-ActionOn_NewPosition}{\code{Magellan::ScreenManager$ActionOn_NewPosition()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ActionOn_New_DataIn">}\href{../../Magellan/html/ScreenManager.html#method-ActionOn_New_DataIn}{\code{Magellan::ScreenManager$ActionOn_New_DataIn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Additional_Server_Funcs">}\href{../../Magellan/html/ScreenManager.html#method-Additional_Server_Funcs}{\code{Magellan::ScreenManager$Additional_Server_Funcs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Change_Current_Pos">}\href{../../Magellan/html/ScreenManager.html#method-Change_Current_Pos}{\code{Magellan::ScreenManager$Change_Current_Pos()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Get_Result">}\href{../../Magellan/html/ScreenManager.html#method-Get_Result}{\code{Magellan::ScreenManager$Get_Result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Global_server">}\href{../../Magellan/html/ScreenManager.html#method-Global_server}{\code{Magellan::ScreenManager$Global_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Reseted">}\href{../../Magellan/html/ScreenManager.html#method-Set_Reseted}{\code{Magellan::ScreenManager$Set_Reseted()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="Set_Skipped">}\href{../../Magellan/html/ScreenManager.html#method-Set_Skipped}{\code{Magellan::ScreenManager$Set_Skipped()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="initialize">}\href{../../Magellan/html/ScreenManager.html#method-initialize}{\code{Magellan::ScreenManager$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="server">}\href{../../Magellan/html/ScreenManager.html#method-server}{\code{Magellan::ScreenManager$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Magellan" data-topic="ScreenManager" data-id="ui">}\href{../../Magellan/html/ScreenManager.html#method-ui}{\code{Magellan::ScreenManager$ui()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ToggleState_Screens"></a>}}
\if{latex}{\out{\hypertarget{method-ToggleState_Screens}{}}}
\subsection{Method \code{ToggleState_Screens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$ToggleState_Screens(cond, range)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A number}

\item{\code{y}}{A number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The sum of \code{x} and \code{y}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{add(1, 1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_All_Skipped"></a>}}
\if{latex}{\out{\hypertarget{method-Set_All_Skipped}{}}}
\subsection{Method \code{Set_All_Skipped()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$Set_All_Skipped()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Discover_Skipped_Steps"></a>}}
\if{latex}{\out{\hypertarget{method-Discover_Skipped_Steps}{}}}
\subsection{Method \code{Discover_Skipped_Steps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$Discover_Skipped_Steps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Set_All_Reset"></a>}}
\if{latex}{\out{\hypertarget{method-Set_All_Reset}{}}}
\subsection{Method \code{Set_All_Reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$Set_All_Reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ValidateCurrentPos"></a>}}
\if{latex}{\out{\hypertarget{method-ValidateCurrentPos}{}}}
\subsection{Method \code{ValidateCurrentPos()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$ValidateCurrentPos()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EncapsulateScreens"></a>}}
\if{latex}{\out{\hypertarget{method-EncapsulateScreens}{}}}
\subsection{Method \code{EncapsulateScreens()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$EncapsulateScreens()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetScreens_ui"></a>}}
\if{latex}{\out{\hypertarget{method-GetScreens_ui}{}}}
\subsection{Method \code{GetScreens_ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$GetScreens_ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
