% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layout.R
\name{process_layout}
\alias{process_layout}
\alias{process_layout_process}
\alias{process_layout_pipeline}
\title{Layout for workflows}
\usage{
process_layout(session, ns, sidebar, content)

process_layout_process(session, ns, sidebar, content)

process_layout_pipeline(session, ns, sidebar, content)
}
\arguments{
\item{session}{The R session parameter}

\item{ns}{xxx}

\item{sidebar}{Shiny components}

\item{content}{Shiny components}
}
\value{
A `div()` containing shiny components
}
\description{
Creates a layout to place the sidebar and the main screen
for pipelines and processes.
}
\examples{
NULL

}
