% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readConfigFile}
\alias{readConfigFile}
\title{Read pipelines configuration files}
\usage{
readConfigFile(path, usermod = "dev")
}
\arguments{
\item{path}{xxx}

\item{usermod}{xxxxx}
}
\description{
xxx
}
\examples{
path <- system.file("workflow/PipelineDemo", package = 'MagellanNTK')
readConfigFile(path)

}
