% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{redBtnClass}
\alias{redBtnClass}
\title{CSS class defining a red button}
\format{
An object of class \code{character} of length 1.
}
\usage{
redBtnClass
}
\description{
CSS class defining a red button
}
\section{Fields}{

\describe{
\item{\code{redBtnClass}}{xxxx}
}}

\keyword{datasets}
