% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_dataset.R
\name{Save_Dataset}
\alias{Save_Dataset}
\alias{Save_Dataset_ui}
\alias{Save_Dataset_server}
\title{Load dataset shiny module}
\usage{
Save_Dataset_ui(id)

Save_Dataset_server(id, data)

Save_Dataset(data)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}
}
\value{
xxxxx

xxxxx
}
\description{
A shiny Module to load a dataset.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(sub_R25)
shiny::runApp(Save_Dataset(sub_R25))

\dontshow{\}) # examplesIf}
}
