% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_dataset.R
\name{Save_Dataset}
\alias{Save_Dataset}
\alias{Save_Dataset_ui}
\alias{Save_Dataset_server}
\title{Load dataset shiny module}
\usage{
Save_Dataset_ui(id)

Save_Dataset_server(id, data)
}
\arguments{
\item{id}{xxx}

\item{data}{xxx}
}
\value{
xxxxx
}
\description{
A shiny Module to load a dataset.
}
