% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{SetHistory}
\alias{SetHistory}
\title{Standardize names}
\usage{
SetHistory(obj.se, history)
}
\arguments{
\item{obj.se}{An instance of the class `SummarizedExperiment`}

\item{history}{A `data.frame()`}
}
\description{
Standardize names
}
\examples{
data(lldata)
history <- GetHistory(lldata, 1)
history <- rbind(history, c('Example', 'Step Ex', 'ex_param', 'Ex'))
lldata[[1]] <- SetHistory(lldata[[1]], history)

}
\author{
Samuel Wieczorek
}
