% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{SetHistory}
\alias{SetHistory}
\title{Get the last validated step before current position.}
\usage{
SetHistory(obj, history)
}
\arguments{
\item{obj}{The dataset managed by MagellanNTK}

\item{history}{A `data.frame()`}
}
\value{
A The dataset managed by MagellanNTK
}
\description{
This function returns the indice of the last validated step before
the current step.
}
\examples{
NULL
}
