% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{shinyInput}
\alias{shinyInput}
\title{function for dynamic inputs in DT
#'description xxx}
\usage{
shinyInput(FUN, id, num, ...)
}
\arguments{
\item{FUN}{xxx}

\item{id}{xxx}

\item{num}{xxx}

\item{...}{xxx}
}
\description{
function for dynamic inputs in DT
#'description xxx
}
