% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_Prostar.R
\name{shinyOutput}
\alias{shinyOutput}
\title{Package version
#'description xxx}
\usage{
shinyOutput(FUN, id, num, ...)
}
\arguments{
\item{FUN}{xxx}

\item{id}{xxx}

\item{num}{xxx}

\item{...}{xxx}
}
\description{
Package version
#'description xxx
}
