% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_wf_files}
\alias{source_wf_files}
\title{Source workflow files}
\usage{
source_wf_files(dirpath, verbose = FALSE)
}
\arguments{
\item{dirpath}{xxx}

\item{verbose}{A `boolean` to indicate whether to turn off (FALSE) or ON (TRUE)
the verbose mode for logs.}
}
\value{
NA
}
\description{
xxx
}
\examples{
path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
source_wf_files(path)

}
