% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_wf_files}
\alias{source_wf_files}
\title{Source workflow files}
\usage{
source_wf_files(dirpath, verbose = FALSE)
}
\arguments{
\item{dirpath}{xxx}

\item{verbose}{A boolean}
}
\description{
xxx
}
\examples{
path <- system.file("extdata/workflow/PipelineA", package = 'MagellanNTK')
source_wf_files(path)

}
