% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_process.R, R/timeline_pipeline.R
\name{timelines}
\alias{timelines}
\alias{timeline_process_ui}
\alias{timeline_process_server}
\alias{timeline_pipeline_ui}
\alias{timeline_pipeline_server}
\title{Timelines}
\usage{
timeline_process_ui(id)

timeline_process_server(id, config, status, position, enabled)

timeline_pipeline_ui(id)

timeline_pipeline_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{config}{An instance of the class `Config`}

\item{status}{A boolean which indicates whether the current status of the 
process ,}

\item{position}{An integer which reflects the current position of the cursor
within the steps.}

\item{enabled}{A vector of booleans with the same length as the number of steps (
See the slot steps in the config object). Each element indicates if the
corresponding step is enabled (TRUE) or (DISABLED)}
}
\value{
NA

A Shiny app
}
\description{
xxx
}
\examples{
NULL
}
