% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_process.R, R/timeline_pipeline.R
\name{timelines}
\alias{timelines}
\alias{timeline_process_ui}
\alias{timeline_process_server}
\alias{timeline_pipeline_ui}
\alias{timeline_pipeline_server}
\title{Timelines}
\usage{
timeline_process_ui(id)

timeline_process_server(id, config, status, position, enabled)

timeline_pipeline_ui(id)

timeline_pipeline_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{A `character()` as the id of the Shiny module}

\item{config}{xxx}

\item{status}{xxx,}

\item{position}{xxx}

\item{enabled}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
NULL

}
