% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_h.R, R/timeline_v.R
\name{timelines}
\alias{timelines}
\alias{timeline_h_ui}
\alias{timeline_h_server}
\alias{timeline_v_ui}
\alias{timeline_v_server}
\title{Timelines modules}
\usage{
timeline_h_ui(id)

timeline_h_server(id, config, status, position, enabled)

timeline_v_ui(id)

timeline_v_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{The `id` of the server}

\item{config}{A static `list` (not `reactiveList`) containing the same 
elements as the process module.}

\item{status}{A `reactive vector` which contain the status (validated,
skipped or undone) of each step of the process module. Its length is equal
to the number of steps.}

\item{position}{A `reactive integer` that reflects the position of the
current (active) step.}

\item{enabled}{A `reactive vector` of length the number of steps and which
indicate whether the step is enabled or disabled.}
}
\value{
NA

NA

NA
}
\description{
Shiny Modules to manage timelines.
}
\examples{
if(interactive()){
  library(shiny)
library(shinyWidgets)
library(shinyjs)

options(shiny.fullstacktrace = TRUE)


config <- Config(
  name = 'PipelineA_Description',
  mode = 'process',
  steps = '',
  mandatory = ''
)



ui <- fluidPage(
  actionButton("prevpos", GlobalSettings$tl_h_prev_icon),
  actionButton("nextpos", GlobalSettings$tl_h_next_icon),
  timeline_h_ui('TLh')
  )



server <- function(input, output){
  
  rv <- reactiveValues(
    direction = 0,
    status = c(0, 1, 0, 0),
    current.pos = 1,
    tl.tags.enabled = c(1, 1, 1, 1),
    position = NULL
  )
  
  
  
  
  observeEvent(input$nextpos,{
    if (rv$current.pos != length(config@steps))
      rv$current.pos <- rv$current.pos + 1
  })
  
  observeEvent(input$prevpos,{
    if (rv$current.pos != 1)
      rv$current.pos <- rv$current.pos - 1
  })
  

  timeline_h_server(id = 'TLh',
    config = config,
    status = reactive({rv$status}),
    position = reactive({rv$current.pos}),
    enabled = reactive({rv$tl.tags.enabled})
  )
  
}


shinyApp(ui, server)
}
if(interactive()){
  library(shiny)
  library(shinyWidgets)
  library(shinyjs)

  options(shiny.fullstacktrace = TRUE)
  

ui <- fluidPage(
  actionButton("prevpos", GlobalSettings$tl_v_prev_icon),
  actionButton("nextpos", GlobalSettings$tl_v_next_icon),
  timeline_v_ui('TLv')
)



server <- function(input, output){
  
  rv <- reactiveValues(
    direction = 0,
    status = c(0, 1, 0, 0),
    current.pos = 1,
    tl.tags.enabled = c(1, 1, 1, 1),
    position = NULL
  )
  
  
  config <- Config(
    name = 'Process1',
    mode = 'process',
    steps = c('Step 1', 'Step 2'),
    mandatory = c(FALSE, TRUE)
  )
  
  
  
  timeline_v_server(id = 'TLv',
    config = config,
    status = reactive({rv$status}),
    position = reactive({rv$current.pos}),
    enabled = reactive({rv$tl.tags.enabled})
  )
  
  observeEvent(input$nextpos,{
    if (rv$current.pos != length(config@steps))
      rv$current.pos <- rv$current.pos + 1
  })
  
  observeEvent(input$prevpos,{
    if (rv$current.pos != 1)
      rv$current.pos <- rv$current.pos - 1
  })
}


shinyApp(ui, server)
}
}
\author{
Samuel Wieczorek
}
