% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_h.R, R/timeline_v.R
\name{timelines}
\alias{timelines}
\alias{timeline_h_ui}
\alias{timeline_h_server}
\alias{timeline_v_ui}
\alias{timeline_v_server}
\title{Timelines modules}
\usage{
timeline_h_ui(id)

timeline_h_server(id, config, status, position, enabled)

timeline_v_ui(id)

timeline_v_server(id, config, status, position, enabled)
}
\arguments{
\item{id}{The `id` of the server}

\item{config}{A static `list` (not `reactiveList`) containing the same 
elements as the process module.}

\item{status}{A `reactive vector` which contain the status (validated,
skipped or undone) of each step of the process module. Its length is equal
to the number of steps.}

\item{position}{A `reactive integer` that reflects the position of the
current (active) step.}

\item{enabled}{A `reactive vector` of length the number of steps and which
indicate whether the step is enabled or disabled.}
}
\value{
NA

NA
}
\description{
Shiny Modules to manage timelines.

xxxx
}
\author{
Samuel Wieczorek
}
