% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{ToggleState_NavBtns}
\alias{ToggleState_NavBtns}
\title{xxx}
\usage{
ToggleState_NavBtns(current.pos, nSteps)
}
\arguments{
\item{current.pos}{xxx}

\item{nSteps}{xxx}
}
\value{
NA
}
\description{
xxx
}
