% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{ToggleState_ResetBtn}
\alias{ToggleState_ResetBtn}
\title{xxx}
\usage{
ToggleState_ResetBtn(cond)
}
\arguments{
\item{cond}{xxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
library(shinyjs)
if (interactive()) {
    ui <- shiny::fluidPage()
    server <- function(input, output, session) {
         ToggleState_ResetBtn(FALSE)
         }
    shiny::shinyApp(ui = ui, server = server)
}

}
