% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{ToggleState_Screens}
\alias{ToggleState_Screens}
\title{xxx}
\usage{
ToggleState_Screens(cond, range, is.enabled, rv)
}
\arguments{
\item{cond}{xxx}

\item{range}{xxx}

\item{is.enabled}{xxx}

\item{rv}{xxx}
}
\value{
xxx
}
\description{
Updates the status of steps in range
}
\examples{
library(shinyjs)
if (interactive()) {
    ui <- shiny::fluidPage()
    server <- function(input, output, session) {
        rv <- reactiveValues(
        steps.status = c(TRUE, TRUE, FALSE))
        observe({ToggleState_Screens(FALSE, c(1,2), TRUE, rv)})
         }
    shiny::shinyApp(ui = ui, server = server)
}

}
