% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{ToggleState_Screens}
\alias{ToggleState_Screens}
\title{xxx}
\usage{
ToggleState_Screens(cond, range, is.enabled, steps.info)
}
\arguments{
\item{cond}{xxx}

\item{range}{xxx}

\item{is.enabled}{xxx}

\item{rv}{xxx}
}
\value{
xxx
}
\description{
Updates the status of steps in range
}
