% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toggleWidget}
\alias{toggleWidget}
\title{xxx}
\usage{
toggleWidget(widget, condition)
}
\arguments{
\item{widget}{xxx}

\item{condition}{xxx}
}
\value{
NA
}
\description{
Returns the date and time in timestamp UNIX format.
}
\examples{
if (interactive()) {
    ui <- shiny::fluidPage(uiOutput('foo'))
    
    server <- function(input, output, session) {
    wgt <- actionButton('foo_btn', 'foo')
    output$foo <- renderUI({toggleWidget(wgt, FALSE)})
         }
    shiny::shinyApp(ui = ui, server = server)
}

}
\author{
Samuel Wieczorek
}
