% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toggleWidget}
\alias{toggleWidget}
\title{Hide/show a widget w.r.t a condition.}
\usage{
toggleWidget(widget, condition)
}
\arguments{
\item{widget}{The id of a `Shiny` widget}

\item{condition}{A `logical(1)` to hide/show the widget.}
}
\value{
NA
}
\description{
Wrapper for the toggleWidget function of the package `shinyjs`
}
\author{
Samuel Wieczorek
}
