% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toggleWidget}
\alias{toggleWidget}
\title{xxx}
\usage{
toggleWidget(widget, condition)
}
\arguments{
\item{widget}{xxx}

\item{condition}{xxx}
}
\value{
NA
}
\description{
Returns the date and time in timestamp UNIX format.
}
\author{
Samuel Wieczorek
}
