% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toggleWidget}
\alias{toggleWidget}
\title{Hide/show a widget w.r.t a condition.}
\usage{
toggleWidget(widget, condition)
}
\arguments{
\item{widget}{A `Shiny` widget}

\item{condition}{A `logical(1)` to hide/show the widget.}
}
\value{
NA
}
\description{
Wrapper for the toggleWidget function of the package `shinyjs`
}
\examples{
if (interactive()) {
    ui <- shiny::fluidPage(uiOutput('foo'))
    
    server <- function(input, output, session) {
    wgt <- actionButton('foo_btn', 'foo')
    output$foo <- renderUI({toggleWidget(wgt, FALSE)})
         }
    shiny::shinyApp(ui, server)
}

}
\author{
Samuel Wieczorek
}
