% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{Update_State_Screens}
\alias{Update_State_Screens}
\title{xxx}
\usage{
Update_State_Screens(
  is.skipped,
  steps.status,
  steps.enabled,
  is.enabled,
  current.pos,
  mandatory
)
}
\arguments{
\item{is.skipped}{xxx}

\item{steps.status}{xxx}

\item{steps.enabled}{xxx}

\item{is.enabled}{xxx}

\item{current.pos}{xxx}

\item{mandatory}{xxx}

\item{xxx}{}
}
\description{
xxx
}
