% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{Update_State_Screens}
\alias{Update_State_Screens}
\title{xxx}
\usage{
Update_State_Screens(is.skipped, is.enabled, steps.info, config, current.pos)
}
\arguments{
\item{is.skipped}{xxx}

\item{is.enabled}{xxx}

\item{rv}{xxx}
}
\description{
xxx
}
