% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_nav_funcs.R
\name{Update_State_Screens}
\alias{Update_State_Screens}
\title{Update State Screens}
\usage{
Update_State_Screens(is.skipped, is.enabled, steps.info, config, current.pos)
}
\arguments{
\item{is.skipped}{xxx}

\item{is.enabled}{xxx}

\item{steps.info}{xxx}

\item{config}{xxx}

\item{current.pos}{xxx}
}
\value{
A vector of enabled tag for the steps.
}
\description{
Updates the tag 'enabled' w.r.t. the actual state of
the module (timeline, validated steps, ...)
}
