% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_core_funcs.R
\name{UpdateStepsStatus}
\alias{UpdateStepsStatus}
\title{Update the status of steps in a pipeline nor process.}
\usage{
UpdateStepsStatus(dataIn, config)
}
\arguments{
\item{dataIn}{An instance of the `SummarizedExperiment` class}

\item{config}{An instance of the `Config` class}
}
\value{
A vector of boolean in which each item indicates the status 
(VALIDATED, UNDONE, SKIPPED) of the corresponding step.
}
\description{
Update the status of steps in a pipeline nor process.
}
\examples{
NULL
}
