% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_view_dataset.R
\name{view_dataset}
\alias{view_dataset}
\alias{view_dataset_ui}
\alias{view_dataset_server}
\title{mod_open_demo_dataset_ui and mod_open_demo_dataset_server}
\usage{
view_dataset_ui(id)

view_dataset_server(id, obj = NULL, ...)

view_dataset(obj, ...)
}
\arguments{
\item{id}{xxx}

\item{obj}{An instance of class xxx}

\item{...}{Additional parameters}
}
\description{
A shiny Module.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shiny::runApp(view_dataset(sub_R25))

\dontshow{\}) # examplesIf}
}
