% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{workflow}
\alias{workflow}
\alias{workflow_ui}
\alias{workflow_server}
\alias{workflowApp}
\title{Shiny example module `Process`}
\usage{
workflow_ui(id)

workflow_server(
  id,
  path = NULL,
  dataIn = reactive({
     NULL
 }),
  tl.layout = NULL,
  usermod = "dev",
  verbose = FALSE
)

workflowApp(
  id,
  path = NULL,
  dataIn = NULL,
  tl.layout = NULL,
  usermod = "dev",
  verbose = FALSE
)
}
\arguments{
\item{id}{xxx}

\item{path}{xxx}

\item{dataIn}{xxx}

\item{tl.layout}{Additional parameters for nav}

\item{usermod}{Available values are 'superdev', 'dev', 'superuser', 'user'}

\item{...}{xxx}
}
\value{
NA
}
\description{
This module contains the configuration information for the corresponding 
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developers who want to create their own 
pipelines nor processes to be managed with `MagellanNTK`.
}
\examples{
\dontrun{
data(sub_R25)

path <- system.file("workflow/PipelineDemo", package = "MagellanNTK")
files <- list.files(file.path(path, 'R'), full.names = TRUE)
for(f in files)
  source(f, local = FALSE, chdir = TRUE)

# Nothing happens when dataIn is NULL
shiny::runApp(run_workflow("PipelineDemo_Process1", dataIn = NULL))

shiny::runApp(run_workflow("PipelineDemo_Process1", dataIn = sub_R25))

shiny::runApp(run_workflow("PipelineDemo_Process1", dataIn = data.frame(), tl.layout = "v"))

shiny::runApp(run_workflow("PipelineDemo", dataIn = sub_R25, tl.layout = c("v", "h")))

shiny::runApp(workflow("PipelineDemo", dataIn = sub_R25, tl.layout = c("v", "h")))
}

\dontrun{
data(lldata)
path <- system.file('workflow/PipelineDemo', package = 'MagellanNTK')

# Nothing happens when dataIn is NULL
shiny::runApp(workflowApp("PipelineDemo_Process1", path, dataIn = NULL))

shiny::runApp(workflowApp("PipelineDemo_Process1", path, dataIn = lldata))

shiny::runApp(workflowApp("PipelineDemo_Process1", dataIn = data.frame())

shiny::runApp(workflowApp("PipelineDemo", path, dataIn = lldata))

shiny::runApp(workflowApp("PipelineB", path, tl.layout = c("v", "h")))

shiny::runApp(workflowApp("PipelineDemo", path, dataIn = sub_R25, 
tl.layout = c("v", "h")))
}

}
\author{
Samuel Wieczorek
}
