% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{workflow}
\alias{workflow}
\alias{workflow_ui}
\alias{workflow_server}
\title{Shiny example module `Process`}
\usage{
workflow_ui(id)

workflow_server(
  id,
  dataIn = reactive({
     NULL
 }),
  tl.layout = NULL,
  mode = "user"
)

workflow(id, dataIn = NULL, tl.layout = NULL, mode = "user")
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{tl.layout}{Additional parameters for nav}

\item{mode}{xxx}

\item{name}{xxx}
}
\value{
NA
}
\description{
This module contains the configuration information for the corresponding 
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
This documentation is for developers who want to create their own 
pipelines nor processes to be managed with `MagellanNTK`.

xxx
}
\examples{
if (interactive()) {
  data(sub_R25)
  
  path <- system.file("extdata/workflow/PipelineA", package = "MagellanNTK")
  files <- list.files(file.path(path, 'R'), full.names = TRUE)
  for(f in files)
    source(f, local = FALSE, chdir = TRUE)

  # Nothing happens when dataIn is NULL
    shiny::runApp(run_workflow("PipelineA_Process1", dataIn = NULL))
  
    shiny::runApp(run_workflow("PipelineA_Process1", dataIn = sub_R25))

    shiny::runApp(run_workflow("PipelineA_Process1", dataIn = data.frame(), tl.layout = "v"))
  
    shiny::runApp(run_workflow("PipelineA", dataIn = sub_R25, tl.layout = c("v", "h")))
    
    shiny::runApp(run_workflow("PipelineB", tl.layout = c("v", "h")))
    
    shiny::runApp(workflow("PipelineA", dataIn = sub_R25, tl.layout = c("v", "h")))
}


if (interactive()) {
data(sub_R25)
shiny::runApp(workflow('PipelineA', sub_R25))
}


}
\author{
Samuel Wieczorek
}
