% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_Excel.R
\name{write.excel}
\alias{write.excel}
\title{This function exports a single data.frame to a Excel file.}
\usage{
write.excel(
  df,
  style = NULL,
  tabname = "foo",
  filename = NULL,
  rownames = FALSE,
  colnames = TRUE
)
}
\arguments{
\item{df}{A data.frame}

\item{style}{xxx}

\item{tabname}{xxx}

\item{filename}{A character string for the name of the Excel}

\item{rownames}{A `logical(0)` xxx}

\item{colnames}{A `logical(0)` xxx}
}
\value{
A Excel file (.xlsx)
}
\description{
xxx
}
\examples{
# Colorize the cells with missing values
data(data_na)
n <- nrow(data_na$array1)*ncol(data_na$array1)
style <- data.frame(matrix(rep('lightgrey', n), nrow=nrow(data_na$array1)))
style[is.na(data_na$array1)] <- 'orange'
write.excel(data_na$array1, style, tabname = 'foo', filename = "test")

}
\author{
Samuel Wieczorek
}
